% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_21.R
\name{a_count_occurrences_ser_ae}
\alias{a_count_occurrences_ser_ae}
\title{Analysis Function to Calculate Count/Fraction of Serious Adverse Event Occurrences}
\usage{
a_count_occurrences_ser_ae(
  df,
  .var,
  .N_col,
  df_denom = NULL,
  denom = c("N_s", "N_col", "n"),
  id_var = "USUBJID",
  arm_var = "ARM"
)
}
\arguments{
\item{df}{(\code{data.frame})\cr data set containing all analysis variables.}

\item{.var}{(\code{string})\cr single variable name that is passed by \code{rtables} when requested
by a statistics function.}

\item{.N_col}{(\code{integer(1)})\cr column-wise N (column count) for the full column being analyzed that is typically
passed by \code{rtables}.}

\item{df_denom}{(\code{data.frame})\cr Full data frame used to calculate denominator subgroup counts
when \code{denom = "N_s"}.}

\item{denom}{(\code{character})\cr Denominator to use to calculate fractions. Can be \code{"N_s"} (total \code{df_denom}
subgroup/row counts), \code{"N_col"} (total \code{df} column counts), or \code{"n"} (total \code{df} overall patient count).
Note that \code{df} is filtered to only include serious adverse events (\code{ASER == "Y"}).}

\item{id_var}{(\code{character})\cr Name of the unique subject identifiers variable.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}
}
\description{
Analysis Function to Calculate Count/Fraction of Serious Adverse Event Occurrences
}
\keyword{internal}
