% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_05.R
\name{a_trtdur_stats}
\alias{a_trtdur_stats}
\title{Analysis Function to Calculate Statistics for Treatment Duration}
\usage{
a_trtdur_stats(x, u_trtdur, .spl_context)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numbers to analyze.}

\item{u_trtdur}{(\code{character})\cr unit for duration of treatment. Options are \code{"days"}, \code{"weeks"}, \code{"months"},
and \code{"years"}.}

\item{.spl_context}{(\code{data.frame})\cr data frame containing information about ancestor split states
that is passed by \code{rtables}.}
}
\description{
Analysis Function to Calculate Statistics for Treatment Duration
}
\keyword{internal}
