% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alt_counts_df_preproc}
\alias{alt_counts_df_preproc}
\title{Preprocess \code{alt_counts_df} for Safety Population}
\usage{
alt_counts_df_preproc(alt_counts_df)
}
\arguments{
\item{alt_counts_df}{(\code{character}) alternative (full) data (typically ADSL) used only to calculate column counts.}
}
\description{
If \code{alt_counts_df} is not \code{NULL}, will check for required variables and filter
to include only safety population, then apply \code{\link[tern:df_explicit_na]{tern::df_explicit_na()}}.
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
alt_counts_df_preproc(adsl)

}
