% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_03.R
\name{ard_make_table_03}
\alias{ard_make_table_03}
\alias{ard_table_03}
\title{Make ARD: Table 3}
\usage{
ard_table_03(
  df,
  denominator = NULL,
  arm_var = "ARM",
  id_var = "USUBJID",
  scrnfl_var = "SCRNFL",
  scrnfailfl_var = "SCRNFAILFL",
  scrnfail_var = "SCRNFRS"
)
}
\description{
Make ARD: Table 3
}
\examples{
library(dplyr)

adsl <- random.cdisc.data::cadsl

set.seed(1)
adsl$RANDDT[sample(seq_len(nrow(adsl)), 100)] <- NA
scrnfail_reas_lvls <- c(
  "Inclusion/exclusion criteria not met", "Patient noncompliance", "Consent withdrawn", "Other"
)
adsl <- adsl \%>\%
  mutate(
    ENRLDT = RANDDT,
    SCRNFL = "Y",
    SCRNFRS = factor(sample(scrnfail_reas_lvls, size = nrow(adsl), replace = TRUE),
      levels = scrnfail_reas_lvls
    ),
    SCRNFAILFL = ifelse(is.na(ENRLDT), "Y", "N")
  )
adsl$SCRNFRS[adsl$SCRNFL == "N" | !is.na(adsl$ENRLDT)] <- NA

ard <- cardinal:::ard_table_03(df = df)
ard

}
\keyword{internal}
