% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_05.R
\name{ard_make_table_05}
\alias{ard_make_table_05}
\alias{ard_table_05}
\title{Make ARD: Table 5}
\usage{
ard_table_05(
  df,
  alt_counts_df = NULL,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  trtsdtm_var = "TRTSDTM",
  trtedtm_var = "TRTEDTM",
  u_trtdur = "days"
)
}
\description{
Make ARD: Table 5
}
\examples{
adsl <- random.cdisc.data::cadsl
df <- preproc_df_table_05(
  adsl,
  saffl_var = "SAFFL",
  trtsdtm_var = "TRTSDTM",
  trtedtm_var = "TRTEDTM",
  u_trtdur = "days"
)

tbl <- ard_table_05(df = df)
tbl

}
\keyword{internal}
