% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_11.R
\name{ard_make_table_11}
\alias{ard_make_table_11}
\alias{ard_table_11}
\title{Make ARD: Table 11}
\usage{
ard_table_11(
  df,
  denominator = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  fmqsc_var = "FMQ01SC",
  fmqnam_var = "FMQ01NAM",
  fmq_scope = "NARROW",
  na_level = "<Missing>"
)
}
\description{
Make ARD: Table 11
}
\examples{
library(dplyr)

adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

set.seed(1)
adae <- adae \%>\%
  rename(FMQ01SC = SMQ01SC) \%>\%
  mutate(
    AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE),
    FMQ01NAM = sample(c("FMQ1", "FMQ2", "FMQ3"), size = nrow(adae), replace = TRUE)
  )
adae$DCSREAS[is.na(adae$DCSREAS)] <- "ADVERSE EVENT"
adae$FMQ01SC[is.na(adae$FMQ01SC)] <- "NARROW"

ard <- cardinal:::ard_table_11(adae, adsl)
ard

}
\keyword{internal}
