% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_12.R
\name{ard_make_table_12}
\alias{ard_make_table_12}
\alias{ard_table_12}
\title{Make ARD: Table 12}
\usage{
ard_table_12(
  df,
  denominator = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  pref_var = "AEDECOD"
)
}
\description{
Make ARD: Table 12
}
\examples{
library(dplyr)

adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

set.seed(1)
adae <- adae \%>\%
  rename(FMQ01SC = SMQ01SC) \%>\%
  mutate(
    AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE),
    FMQ01NAM = sample(c("FMQ1", "FMQ2", "FMQ3"), size = nrow(adae), replace = TRUE)
  )
adae$DCSREAS[is.na(adae$DCSREAS)] <- "ADVERSE EVENT"
adae$FMQ01SC[is.na(adae$FMQ01SC)] <- "NARROW"

ard <- cardinal:::ard_table_12(adae, adsl)
ard

}
\keyword{internal}
