% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_32.R
\name{ard_make_table_32}
\alias{ard_make_table_32}
\alias{ard_table_32}
\title{Make ARD: Table 32}
\usage{
ard_table_32(
  df,
  alt_counts_df = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  subset,
  lbl_overall = NULL
)
}
\description{
Make ARD: Table 32
}
\examples{
adsl <- random.cdisc.data::cadsl
advs <- random.cdisc.data::cadvs

ard <- cardinal:::ard_table_32(
  df = advs,
  alt_counts_df = adsl,
  subset = (PARAMCD == "DIABP" & AVISITN >= 1)
)
ard

}
\keyword{internal}
