% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_35.R
\name{ard_make_table_35}
\alias{ard_make_table_35}
\alias{ard_table_35}
\title{Make ARD: Table 35}
\usage{
ard_table_35(
  df,
  denominator = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  soc_var = "AEBODSYS",
  lbl_overall = NULL,
  na_level = "<Missing>"
)
}
\description{
Make ARD: Table 35
}
\examples{
library(dplyr)

adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

ard <- cardinal:::ard_table_35(adae, adsl)
ard

}
\keyword{internal}
