% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_05.R
\name{ard_table_05}
\alias{ard_table_05}
\title{ARD Table 05}
\usage{
ard_table_05(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  id_var = "USUBJID",
  saffl_var = "SAFFL",
  trtsdtm_var = "TRTSDTM",
  trtedtm_var = "TRTEDTM",
  u_trtdur = "days",
  lbl_trtdur = paste("Duration of Treatment,", u_trtdur),
  lbl_overall = NULL,
  risk_diff = NULL,
  prune_0 = FALSE,
  annotations = NULL
)
}
\description{
ARD Table 05
}
\examples{
adsl <- random.cdisc.data::cadsl
tbl <- ard_table_05(df = adsl)

tbl

}
