% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{argument_convention}
\alias{argument_convention}
\title{Standard Arguments}
\arguments{
\item{show_colcounts}{(\code{flag})\cr Whether column counts should be printed. Boolean.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{id_var}{(\code{character})\cr Name of the unique subject identifiers variable.}

\item{saffl_var}{(\code{character})\cr Name of the safety flag variable which must contain at least "Y" for "Yes".}

\item{ser_var}{(\code{character})\cr Name of the serious adverse event flag variable which must contain at least
"Y" for "Yes".}

\item{lbl_overall}{(\code{character})\cr Optional. If specified, an overall column will be added to the table with
the given value as the column label.}

\item{vars}{(\code{vector} of \code{character})\cr Variables from \code{df} to include in the table.}

\item{lbl_vars}{(\code{vector} of \code{character})\cr Labels corresponding to variables in \code{vars} to print
in the table. Labels should be ordered according to the order of variables in \code{vars}.}

\item{pref_var}{(\code{character})\cr Name of the preferred term variable from \code{adae} to include in the table.}

\item{soc_var}{(\code{character})\cr Name of the system organ class variable from \code{adae} to include in the table.}

\item{lbl_pref_var}{(\code{character})\cr Label corresponding to preferred term variable \code{pref_var} to print in the table.}

\item{lbl_soc_var}{(\code{character})\cr Label corresponding to system organ class variable \code{soc_var} to print in the
table.}

\item{.stats}{(\code{character})\cr Statistics to include in the table. Includes statistics for all variable
types (only the statistics that are valid for a given variable's type will be printed).
See \code{\link[tern:analyze_vars]{tern::analyze_vars()}} for options.}

\item{.formats}{(named \code{list} of \code{character})\cr List of formats corresponding to each value in \code{.stats}.
Each name is a value in \code{.stats} and the corresponding value is the format that should be applied to
that statistic. See \code{\link[formatters:list_valid_format_labels]{formatters::list_valid_format_labels()}} for a list of valid formats.}

\item{na_rm}{(\code{flag})\cr Whether \code{NA} levels should be removed from the table.}

\item{prune_0}{(\code{flag})\cr Whether all-zero rows should be removed from the table. Boolean.}

\item{na_level}{(\code{character})\cr String to represent missing values.}

\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{adex}{(\code{data.frame})\cr dataset (typically ADEX) required to build table.}

\item{advs}{(\code{data.frame})\cr dataset (typically ADVS) required to build table.}

\item{adsl}{(\code{data.frame})\cr dataset (typically ADSL) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}

\item{ard}{(\code{cards::card})\cr an ARD.}

\item{denominator}{(\code{data.frame})\cr alternative dataset (typically ADSL) used only to calculate denominator counts.}

\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{eosdy_var}{(\code{character})\cr variable denoting last recorded (relative) study day.}

\item{fmqsc_var}{(\code{character})\cr FMQ scope variable to use in table.}

\item{fmqnam_var}{(\code{character})\cr FMQ reference name variable to use in table.}

\item{fmq_scope}{(\code{character})\cr FMQ scope ("NARROW" or "BROAD") to output in table.}

\item{ggtheme}{(\code{theme})\cr a graphical theme as provided by \code{ggplot2} to control styling of the \code{ggplot} object.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add one or more risk difference columns to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the names of the arms to compare to the reference arm. A new column will be added for each
element of \code{arm_y}.
\item \code{col_label}: (optional) labels to use for the risk difference columns. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}. For more than one risk difference column, \code{"arm x vs. arm y"} text will also
be included in the column labels by default. The length of \code{col_label} must be equal to the length of \code{arm_y}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}

\item{sex_scope}{(\code{character})\cr Level of \code{SEX} to output in table.}

\item{table_engine}{(\code{string})\cr table engine used to generate the table. If \code{NULL}, no table is returned. If the
given engine is not available for the table, no table is created and a warning is returned.}

\item{trtedtm_var}{(\code{character})\cr treatment end datetime variable.}

\item{trtemfl_var}{(\code{character})\cr flag variable used to identify Treatment-emergent AE.}

\item{trtsdtm_var}{(\code{character})\cr treatment start datetime variable.}

\item{u_trtdur}{(\code{character})\cr unit for duration of treatment. Options are \code{"days"}, \code{"weeks"}, \code{"months"},
and \code{"years"}.}

\item{xticks}{(\code{vector} of \code{numeric})\cr x-axis tick positions. If \code{NA} (default), tick mark positions are
automatically calculated.}

\item{x_lab}{(\code{character})\cr x-axis label.}

\item{yticks}{(\code{vector} of \code{numeric})\cr y-axis tick positions. If \code{NA} (default), tick mark positions are
automatically calculated.}

\item{y_lab}{(\code{character})\cr y-axis label.}
}
\description{
This documentation lists all the arguments in \code{cardinalfda} that are used repeatedly by functions to output tables.
}
\keyword{internal}
