% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{basic_table_annot}
\alias{basic_table_annot}
\title{Initialize Table Layout with Annotations}
\usage{
basic_table_annot(show_colcounts = TRUE, annotations = NULL)
}
\arguments{
\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
Initializes a table layout via \code{\link[rtables:basic_table]{rtables::basic_table()}} and applies \code{show_colcounts}
argument. Adds any annotations supplied via the \code{annotations} argument.
}
\examples{
lyt <- basic_table_annot(annotations = list(title = "Title", main_footer = "Main Footer"))
rtables::build_table(lyt, df = data.frame())

}
