% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R, R/fda-table_10.R
\name{calculate_riskdiff}
\alias{calculate_riskdiff}
\title{helper function to calculate the risk difference for table 9}
\usage{
calculate_riskdiff(x, y, n_x, n_y)

calculate_riskdiff(x, y, n_x, n_y)
}
\arguments{
\item{x}{character vector containing all value for the first treatment}

\item{y}{character vector containing all value for the second treatment}

\item{n_x}{character vector containing for all values of x the corresponding N}

\item{n_y}{character vector containing for all values of y the corresponding N}
}
\value{
vector of characters containing the value for risk difference and
the corresponding 95\% Confidence interval in Brackets

vector of characters containing the value for risk difference and
the corresponding 95\% Confidence interval in Brackets
}
\description{
helper function to calculate the risk difference for table 9

helper function to calculate the risk difference for table 10
}
