% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{count_subjects}
\alias{count_subjects}
\title{Helper function for \code{create_table_09_data()}
Used for counting subjects per group}
\usage{
count_subjects(adae, arm_var, sub_level_vars = NULL, lbl_overall = NULL)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{sub_level_vars}{(\code{NULL} or \code{character}) specifying the sub group for counted subjects}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}
}
\value{
A \code{data.frame} containing the number of subjects with \code{sub_level_vars} events
}
\description{
Helper function for \code{create_table_09_data()}
Used for counting subjects per group
}
