% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_10.R
\name{create_table_10_data}
\alias{create_table_10_data}
\title{Helper function to create the data for \code{make_table_10_gt()}}
\usage{
create_table_10_data(
  adae,
  alt_counts_df,
  arm_var,
  id_var,
  soc_var,
  fmqsc_var,
  fmqnam_var,
  lbl_overall = NULL,
  risk_diff = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{soc_var}{(\code{character})\cr flag variable used to indicate system organ class.}

\item{fmqsc_var}{(\code{character})\cr FMQ scope variable to use in table.}

\item{fmqnam_var}{(\code{character})\cr FMQ reference name variable to use in table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add a risk difference column to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the name of the arm to compare to the reference arm.
\item \code{col_label}: (optional) label to use for the risk difference column. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}
}
\value{
list containing the counted data to be displayed for table 10 and
a \code{data.frame} containing information about the total N for each group
}
\description{
Helper function to create the data for \code{make_table_10_gt()}
}
\details{
If \code{lbl_overall} is not \code{NULL} only the data for the overall column will be generated
}
