\name{defunct}
\alias{Defunct}

\alias{generateImage}
\alias{generateSpectrum}

\title{Defunct functions and methods in Cardinal}

\description{
  These functions are defunct and are no longer available.

  \describe{
        \item{\code{generateSpectrum}:}{Use \code{\link{simulateSpectrum}} instead.}
        \item{\code{generateImage}:}{Use \code{\link{simulateImage}} instead.}
    }
}
