% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon.R
\docType{package}
\name{falcon-package}
\alias{falcon}
\alias{falcon-package}
\title{falcon Package}
\description{
Implementation of FDA Safety Tables and Figures
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmaverse/falcon/}
  \item Report bugs at \url{https://github.com/pharmaverse/falcon/issues}
}

}
\author{
\strong{Maintainer}: Pawel Rucki \email{pawel.rucki@roche.com}

}
\keyword{internal}
