% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{h_make_table_09}
\alias{h_make_table_09}
\alias{create_table_09_data}
\alias{count_subjects}
\alias{renaming_function}
\alias{calculate_riskdiff}
\title{Helper functions used to construct FDA table 9}
\usage{
create_table_09_data(
  adae,
  alt_counts_df,
  arm_var,
  id_var,
  soc_var,
  pref_var,
  lbl_overall = NULL,
  risk_diff = NULL
)

count_subjects(
  adae,
  arm_var,
  id_var,
  sub_level_vars = NULL,
  lbl_overall = NULL,
  risk_diff = NULL
)

renaming_function(x)

calculate_riskdiff(x, y, n_x, n_y)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{soc_var}{(\code{character})\cr Name of the system organ class variable from \code{adae} to include in the table.}

\item{pref_var}{(\code{character})\cr Name of the preferred term variable from \code{adae} to include in the table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add one or more risk difference columns to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the names of the arms to compare to the reference arm. A new column will be added for each
element of \code{arm_y}.
\item \code{col_label}: (optional) labels to use for the risk difference columns. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}. For more than one risk difference column, \code{"arm x vs. arm y"} text will also
be included in the column labels by default. The length of \code{col_label} must be equal to the length of \code{arm_y}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}

\item{sub_level_vars}{(\code{character} or \code{NULL}) the subgroup for counted subjects.}

\item{x}{(\code{character})\cr vector containing all values for the first treatment.}

\item{y}{(\code{character})\cr vector containing all values for the second treatment.}

\item{n_x}{(\code{character})\cr vector containing for all values of \code{x} the corresponding N.}

\item{n_y}{(\code{character})\cr vector containing for all values of \code{y} the corresponding N.}
}
\value{
\itemize{
\item \code{create_table_09_data} returns a list containing the counted data to be displayed for table 9 and
a \code{data.frame} containing information about the total N for each group.
}

\itemize{
\item \code{count_subjects} returns a \code{data.frame} containing the number of subjects with \code{sub_level_vars} events.
}

\itemize{
\item \code{renaming_function} returns a vector.
}

\itemize{
\item \code{calculate_riskdiff} returns a vector of characters containing the value for risk difference and
the corresponding 95\% confidence interval.
}
}
\description{
Helper functions used to construct FDA table 9
}
\details{
\itemize{
\item \code{create_table_09_data}: If \code{lbl_overall} is non-\code{NULL}, only the data for the overall column will be generated.
}
}
\section{Functions}{
\itemize{
\item \code{create_table_09_data()}: Helper function to create the data for \code{\link[=make_table_09_gtsum]{make_table_09_gtsum()}}.

\item \code{count_subjects()}: Helper function used to count subjects per group.

\item \code{renaming_function()}: Helper function to rename the combined columns.

\item \code{calculate_riskdiff()}: Helper function to calculate the risk difference for table 9.

}}
\seealso{
\code{\link[=make_table_09]{make_table_09()}}, \code{\link[=make_table_09_tplyr]{make_table_09_tplyr()}}, \code{\link[=make_table_09_gtsum]{make_table_09_gtsum()}}
}
\keyword{internal}
