\name{internal}
\alias{Cardinal-internal}

\alias{coregister}
\alias{coregister-methods}
\alias{coregister,SpatialKMeans,missing-method}
\alias{coregister,SpatialShrunkenCentroids,missing-method}

\alias{class:SummaryDataFrame}
\alias{SummaryDataFrame}
\alias{SummaryDataFrame-class}

\alias{show,SummaryDataFrame-method}
\alias{summary,SummaryDataFrame-method}

\title{Internal utilities for Cardinal}

\description{
	Low-level utility functions, classes, and data defined in the \pkg{Cardinal} package. They are not intended to be used directly.
}

\author{Kylie A. Bemis}

\keyword{utilities}
\keyword{internal}
