\name{legacy}
\alias{Cardinal-legacy}

\alias{batchProcess}
\alias{batchProcess-methods}
\alias{batchProcess,MSImageSet-method}

\alias{class:Hashmat}
\alias{Hashmat}
\alias{Hashmat-class}

\alias{[,Hashmat-method}
\alias{[<-,Hashmat-method}
\alias{[,Hashmat,ANY,ANY,ANY-method}
\alias{[,Hashmat,ANY,ANY,NULL-method}
\alias{[<-,Hashmat,ANY,ANY,ANY-method}
\alias{dim,Hashmat-method}
\alias{dim<-,Hashmat-method}
\alias{dimnames,Hashmat-method}
\alias{dimnames<-,Hashmat-method}
\alias{dimnames<-,Hashmat,ANY-method}
\alias{colnames,Hashmat-method}
\alias{colnames<-,Hashmat-method}
\alias{rownames,Hashmat-method}
\alias{rownames<-,Hashmat-method}
\alias{cbind,Hashmat-method}
\alias{rbind,Hashmat-method}
\alias{pData,Hashmat-method}
\alias{pData<-,Hashmat-method}
\alias{pData<-,Hashmat,ANY-method}
\alias{keys}
\alias{keys<-}
\alias{keys,Hashmat-method}
\alias{keys<-,Hashmat-method}
\alias{keys<-,Hashmat,character-method}
\alias{keys<-,Hashmat,list-method}
\alias{combine,Hashmat,Hashmat-method}
\alias{initialize,Hashmat-method}
\alias{show,Hashmat-method}

\alias{class:IAnnotatedDataFrame}
\alias{IAnnotatedDataFrame}
\alias{IAnnotatedDataFrame-class}

\alias{coord,IAnnotatedDataFrame-method}
\alias{coord<-,IAnnotatedDataFrame-method}
\alias{coordLabels,IAnnotatedDataFrame-method}
\alias{coordLabels<-,IAnnotatedDataFrame-method}
\alias{initialize,IAnnotatedDataFrame-method}
\alias{pixelNames,IAnnotatedDataFrame-method}
\alias{pixelNames<-,IAnnotatedDataFrame-method}
\alias{sampleNames,IAnnotatedDataFrame-method}
\alias{sampleNames<-,IAnnotatedDataFrame-method}
\alias{sampleNames<-,IAnnotatedDataFrame,ANY-method}
\alias{combine,IAnnotatedDataFrame,IAnnotatedDataFrame-method}
\alias{[,IAnnotatedDataFrame,ANY,ANY,ANY-method}
\alias{is3D,IAnnotatedDataFrame-method}
\alias{show,IAnnotatedDataFrame-method}

\alias{class:ImageData}
\alias{ImageData}
\alias{ImageData-class}

\alias{[[,ImageData-method}
\alias{[[<-,ImageData-method}
\alias{[[,ImageData,character,missing-method}
\alias{[[<-,ImageData,character,missing-method}
\alias{annotatedDataFrameFrom,ImageData-method}
\alias{combine,ImageData,ImageData-method}
\alias{combine,vector,vector-method}
\alias{combine,array,array-method}
\alias{dim,ImageData-method}
\alias{dims,ImageData-method}
\alias{initialize,ImageData-method}
\alias{storageMode,ImageData-method}
\alias{storageMode<-,ImageData,character-method}
\alias{names,ImageData-method}
\alias{names<-,ImageData-method}
\alias{show,ImageData-method}

\alias{class:iSet}
\alias{iSet}
\alias{iSet-class}

\alias{$,iSet-method}
\alias{$<-,iSet-method}
\alias{[[,iSet-method}
\alias{[[<-,iSet-method}
\alias{[[,iSet,ANY,ANY-method}
\alias{[[<-,iSet,ANY,ANY-method}
\alias{initialize,iSet-method}
\alias{length,iSet-method}
\alias{dim,iSet-method}
\alias{dims,iSet-method}
\alias{storageMode,iSet-method}
\alias{storageMode<-,iSet,character-method}
\alias{storageMode<-,iSet,ANY-method}
\alias{combine,iSet,iSet-method}
\alias{protocolData,iSet-method}
\alias{protocolData<-,iSet,ANY-method}
\alias{experimentData,iSet-method}
\alias{experimentData<-,iSet,ANY-method}
\alias{imageData,iSet-method}
\alias{imageData<-,iSet-method}
\alias{iData,iSet-method}
\alias{iData<-,iSet-method}
\alias{pixels,iSet-method}
\alias{pixelData,iSet-method}
\alias{pixelData<-,iSet-method}
\alias{pData,iSet-method}
\alias{pData<-,iSet-method}
\alias{pData<-,iSet,ANY-method}
\alias{varMetadata,iSet-method}
\alias{varMetadata<-,iSet-method}
\alias{varMetadata<-,iSet,ANY-method}
\alias{varLabels,iSet-method}
\alias{varLabels<-,iSet-method}
\alias{sampleNames,iSet-method}
\alias{sampleNames<-,iSet-method}
\alias{sampleNames<-,iSet,ANY-method}
\alias{pixelNames,iSet-method}
\alias{pixelNames<-,iSet-method}
\alias{coordLabels,iSet-method}
\alias{coordLabels<-,iSet-method}
\alias{coord,iSet-method}
\alias{coord<-,iSet-method}
\alias{features,iSet-method}
\alias{featureData,iSet-method}
\alias{featureData<-,iSet-method}
\alias{featureData<-,iSet,ANY-method}
\alias{fData,iSet-method}
\alias{fData<-,iSet-method}
\alias{fData<-,iSet,ANY-method}
\alias{fvarMetadata,iSet-method}
\alias{fvarMetadata<-,iSet-method}
\alias{fvarMetadata<-,iSet,ANY-method}
\alias{fvarLabels,iSet-method}
\alias{fvarLabels<-,iSet-method}
\alias{featureNames,iSet-method}
\alias{featureNames<-,iSet-method}
\alias{summary,iSet-method}
\alias{print.summary.iSet}
\alias{show,iSet-method}

\alias{class:MIAPE-Imaging}
\alias{MIAPE-Imaging}
\alias{MIAPE-Imaging-class}

\alias{show,MIAPE-Imaging-method}
\alias{abstract,MIAPE-Imaging-method}
\alias{samples,MIAPE-Imaging-method}
\alias{pubMedIds,MIAPE-Imaging-method}
\alias{pubMedIds<-,MIAPE-Imaging,ANY-method}
\alias{otherInfo,MIAPE-Imaging-method}
\alias{expinfo,MIAPE-Imaging-method}
\alias{notes,MIAPE-Imaging-method}
\alias{notes<-,MIAPE-Imaging,character-method}
\alias{notes<-,MIAPE-Imaging,list-method}
\alias{preproc,MIAPE-Imaging-method}
\alias{preproc<-,MIAPE-Imaging-method}
\alias{combine,MIAPE-Imaging,MIAPE-Imaging-method}

\alias{msiInfo,MIAPE-Imaging-method}
\alias{specimenOrigin}
\alias{specimenOrigin,MIAPE-Imaging-method}
\alias{specimenType}
\alias{specimenType,MIAPE-Imaging-method}
\alias{stainingMethod}
\alias{stainingMethod,MIAPE-Imaging-method}
\alias{tissueThickness}
\alias{tissueThickness,MIAPE-Imaging-method}
\alias{tissueWash}
\alias{tissueWash,MIAPE-Imaging-method}
\alias{embeddingMethod}
\alias{embeddingMethod,MIAPE-Imaging-method}
\alias{inSituChemistry}
\alias{inSituChemistry,MIAPE-Imaging-method}

\alias{matrixApplication,MIAPE-Imaging-method}
\alias{pixelSize,MIAPE-Imaging-method}
\alias{instrumentModel,MIAPE-Imaging-method}
\alias{instrumentVendor,MIAPE-Imaging-method}
\alias{massAnalyzerType,MIAPE-Imaging-method}
\alias{ionizationType,MIAPE-Imaging-method}
\alias{scanPolarity,MIAPE-Imaging-method}
\alias{softwareName}
\alias{softwareName,MIAPE-Imaging-method}
\alias{softwareVersion}
\alias{softwareVersion,MIAPE-Imaging-method}

\alias{scanType,MIAPE-Imaging-method}
\alias{scanPattern,MIAPE-Imaging-method}
\alias{scanDirection,MIAPE-Imaging-method}
\alias{lineScanDirection,MIAPE-Imaging-method}

\alias{imageShape}
\alias{imageShape,MIAPE-Imaging-method}

\alias{class:MSImageData}
\alias{MSImageData}
\alias{MSImageData-class}

\alias{initialize,MSImageData-method}
\alias{peakData,MSImageData-method}
\alias{peakData<-,MSImageData-method}
\alias{mzData,MSImageData-method}
\alias{mzData<-,MSImageData-method}

\alias{class:MSImageProcess}
\alias{MSImageProcess}
\alias{MSImageProcess-class}

\alias{show,MSImageProcess-method}
\alias{combine,MSImageProcess,MSImageProcess-method}
\alias{initialize,MSImageProcess-method}
\alias{normalization,MSImageProcess-method}
\alias{normalization<-,MSImageProcess-method}
\alias{smoothing,MSImageProcess-method}
\alias{smoothing<-,MSImageProcess-method}
\alias{baselineReduction,MSImageProcess-method}
\alias{baselineReduction<-,MSImageProcess-method}
\alias{spectrumRepresentation,MSImageProcess-method}
\alias{spectrumRepresentation<-,MSImageProcess-method}
\alias{peakPicking,MSImageProcess-method}
\alias{peakPicking<-,MSImageProcess-method}
\alias{centroided,MSImageProcess-method}
\alias{centroided<-,MSImageProcess-method}

\alias{class:MSImageSet}
\alias{MSImageSet}
\alias{MSImageSet-class}

\alias{initialize,MSImageSet-method}
\alias{combine,MSImageSet,MSImageSet-method}
\alias{mz,MSImageSet-method}
\alias{mz<-,MSImageSet-method}
\alias{spectra,MSImageSet-method}
\alias{spectra<-,MSImageSet-method}
\alias{peaks,MSImageSet-method}
\alias{peaks<-,MSImageSet-method}
\alias{processingData,MSImageSet-method}
\alias{processingData<-,MSImageSet-method}
\alias{centroided,MSImageSet-method}
\alias{centroided<-,MSImageSet-method}
\alias{pixels,MSImageSet-method}
\alias{features,MSImageSet-method}
\alias{show,MSImageSet-method}

\alias{processingData}
\alias{processingData<-}
\alias{processingData-methods}

\alias{reduceDimension}
\alias{reduceDimension-methods}
\alias{reduceDimension,MSImageSet,missing-method}
\alias{reduceDimension,MSImageSet,numeric-method}
\alias{reduceDimension,MSImageSet,MSImageSet-method}

\alias{reduceDimension.bin}
\alias{reduceDimension.resample}
\alias{reduceDimension.peaks}

\alias{class:ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}

\alias{$,ResultSet-method}
\alias{[,ResultSet-method}
\alias{[,ResultSet,ANY,ANY,ANY-method}
\alias{[[,ResultSet-method}
\alias{[[,ResultSet,ANY,ANY-method}
\alias{resultData,ResultSet-method}
\alias{resultData<-,ResultSet-method}
\alias{resultData,ResultSet,ANY-method}
\alias{resultData<-,ResultSet,missing-method}
\alias{modelData,ResultSet-method}
\alias{modelData<-,ResultSet-method}
\alias{length,ResultSet-method}
\alias{names,ResultSet-method}
\alias{show,ResultSet-method}

\alias{class:SImageData}
\alias{SImageData}
\alias{SImageData-class}

\alias{[,SImageData-method}
\alias{[,SImageData,ANY,ANY,ANY-method}
\alias{[,SImageData,ANY,ANY,NULL-method}
\alias{combine,SImageData,SImageData-method}
\alias{dim,SImageData-method}
\alias{dims,SImageData-method}
\alias{initialize,SImageData-method}
\alias{iData,SImageData,ANY-method}
\alias{iData<-,SImageData,ANY-method}
\alias{peakData,SImageData-method}
\alias{peakData<-,SImageData-method}
\alias{mzData,SImageData-method}
\alias{mzData<-,SImageData-method}
\alias{coord,SImageData-method}
\alias{coord<-,SImageData-method}
\alias{featureNames,SImageData-method}
\alias{featureNames<-,SImageData-method}
\alias{pixelNames,SImageData-method}
\alias{pixelNames<-,SImageData-method}
\alias{positionArray}
\alias{positionArray<-}
\alias{positionArray,SImageData-method}
\alias{positionArray<-,SImageData-method}
\alias{regeneratePositions}
\alias{regeneratePositions,SImageData-method}

\alias{class:SImageSet}
\alias{SImageSet}
\alias{SImageSet-class}

\alias{[,SImageSet-method}
\alias{[,SImageSet,ANY,ANY,ANY-method}
\alias{initialize,SImageSet-method}
\alias{combine,SImageSet,SImageSet-method}
\alias{iData,SImageSet,ANY-method}
\alias{iData<-,SImageSet,ANY-method}
\alias{regeneratePositions,SImageSet-method}
\alias{coord<-,SImageSet-method}
\alias{coordLabels<-,SImageSet-method}
\alias{featureNames<-,SImageSet-method}
\alias{pixelNames<-,SImageSet-method}
\alias{is3D,SImageSet-method}

\alias{standardizeRuns}
\alias{standardizeRuns-methods}
\alias{standardizeRuns,MSImageSet-method}

\alias{standardizeRuns.sum}

\title{Legacy classes and methods in Cardinal}

\description{
  These classes and methods are deprecated and should no longer be used.

  The class definitions will remain for compatibility with \code{CardinalWorkflows} datasets and other older serialized datasets.

  Objects of these classes should be updated to a supported version of the class.

  For example, for an object of class \code{MSImageSet}:

  \code{object <- as(object, "MSImagingExperiment")}
}
