% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-fig_01.R
\name{make_fig_1}
\alias{make_fig_1}
\title{FDA Figure 1: Time to Permanent Discontinuation of Study Drug, Safety Population, Pooled Analyses}
\usage{
make_fig_1(
  df,
  alt_counts_df = NULL,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  trtsdtm_var = "TRTSDTM",
  trtedtm_var = "TRTEDTM",
  u_trtdur = "days",
  x_lab = paste0("Time from first dose (", u_trtdur, ")"),
  y_lab = "Percent of Patients",
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{trtsdtm_var}{(\code{character})\cr treatment start datetime variable in \code{df}.}

\item{trtedtm_var}{(\code{character})\cr treatment end datetime variable in \code{df}.}

\item{u_trtdur}{(\code{character})\cr unit for duration of treatment. Options are \code{"days"}, \code{"weeks"}, \code{"months"},
and \code{"years"}.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\value{
A \code{ggplot2} object.
}
\description{
FDA Figure 1: Time to Permanent Discontinuation of Study Drug, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{df} must contain the variables specified by \code{arm_var}, \code{id_var}, \code{saffl_var}, \code{id_var}, \code{trtsdtm_var},
and \code{trtedtm_var}.
\item If specified, \code{alt_counts_df} must contain the variables specified by \code{arm_var}, \code{id_var} and \code{saffl_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false,). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table "Patients Treated" section are the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are not removed by default (see \code{prune_0} argument).
\item Records with missing treatment start and/or end datetime are excluded from all calculations.
}
}
\examples{
adsl <- random.cdisc.data::cadsl

fig <- make_fig_1(df = adsl)
fig

}
