% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_06.R
\name{make_table_06_gtsummary}
\alias{make_table_06_gtsummary}
\title{Table engine for FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses}
\usage{
make_table_06_gtsummary(
  df,
  denominator,
  id_var = "USUBJID",
  return_ard = TRUE,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  sae_var = "AESER",
  sae_cat_vars = list(`SAEs with fatal outcome` = "AESDTH", `Life-threatening SAEs` =
    "AESLIFE", `SAEs requiring hospitalization` = "AESHOSP",
    `SAEs resulting in substantial disruption of normal life functions` = "AESDISAB",
    `Congenital anomaly or birth defect` = "AESCONG", Other = "AESMIE"),
  disc_var = "AEACNOTH",
  dose_mod_var = "AEACN",
  dose_mod_cat_labels = list(`DRUG INTERRUPTED` =
    "AE leading to interruption of study drug", `DOSE REDUCED` =
    "AE leading to reduction of study drug", `DOSE RATE REDUCED` =
    "AE leading to dose delay of study drug", `DOSE INCREASED` = "Other"),
  sev_var = "AESEV"
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{denominator}{(\code{data.frame})\cr alternative dataset (typically ADSL) used only to calculate denominator counts.}

\item{id_var}{(\code{character})\cr Name of the unique subject identifiers variable.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr Name of the safety flag variable which must contain at least "Y" for "Yes".}

\item{sae_var}{(\code{character})\cr Name of the flag variable indicating whether an adverse event is serious ("Y") or
not ("N"). Defaults to AESER from an ADAE ADaM dataset.}

\item{sae_cat_vars}{(named \code{list} of \code{character}) \cr List of names of flag variables for categories of serious
adverse events. These variables may contain "Y" or "No". Defaults to AESDTH, AESLIFE, AESHOSP, AESDISAB, AESCONG,
and AESMIE from an ADAE ADaM dataset. Provide labels through names of the list elements.}

\item{disc_var}{(\code{character}) \cr Name of the variable containing information whether a subject discontinued from
the study as result of an adverse event. Defaults to AEACNOTH from an ADAE ADaM dataset.}

\item{dose_mod_var}{(\code{character}) \cr Name of the variable containing information about dose modification for a
subject due to an adverse event. Defaults to AEACN from an ADAE ADaM dataset.}

\item{dose_mod_cat_labels}{(named \code{list} of \code{character}) \cr List of labels for the dose modification categories as
listed in the details section.}

\item{sev_var}{Name of the variable containing information about the severity of an adverse event. Defaults to AESEV
from an ADAE ADaM dataset.}
}
\value{
A \code{gtsummary} table
}
\description{
Table engine for FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses
}
