% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{make_table_09_gt}
\alias{make_table_09_gt}
\title{Function to create FDA Table 9: Patients With Serious Adverse Events
by System Organ Class and Preferred Term,
Safety Population, Pooled Analyses with /code{gt}}
\usage{
make_table_09_gt(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  saffl_var = "SAFFL",
  ser_var = "AESER",
  arm_var = "ARM",
  pref_var = "AEDECOD",
  id_var = "USUBJID",
  soc_var = "AESOC",
  lbl_pref_var = formatters::var_labels(adae, fill = TRUE)[pref_var],
  lbl_soc_var = formatters::var_labels(adae, fill = TRUE)[soc_var],
  lbl_overall = NULL,
  annotations = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{saffl_var}{(\code{character})\cr safety population flag variable from \code{adae} to include in the table.}

\item{ser_var}{(\code{character})\cr serious Event variable from \code{adae} to include in the table.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{pref_var}{(\code{character})\cr preferred term variable from \code{adae} to include in the table.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{soc_var}{(\code{character})\cr system organ class variable from \code{adae} to include in the table.}

\item{lbl_pref_var}{(\code{character})\cr label corresponding to preferred term variable \code{pref_var} to print in the table.}

\item{lbl_soc_var}{(\code{character})\cr label corresponding to system organ class variable \code{soc_var} to print in the table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, and a list of characters called \code{footnotes}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\value{
A gt table object
}
\description{
Function to create FDA Table 9: Patients With Serious Adverse Events
by System Organ Class and Preferred Term,
Safety Population, Pooled Analyses with /code{gt}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

tbl <- make_table_09(adae = adae, alt_counts_df = adsl)
tbl

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

annotations <- list(
  title = "Table 9. Patients With Serious Adverse Events",
  subtitle = "by System Organ Class and Preferred Term, Safety Population, Pooled Analyses",
  footnotes = list(
    "Duration = [e.g., X week double-blind treatment period or median and a range indicating pooled trial durations].",
    "Difference is shown between [treatment arms] (e.g., difference is shown between Drug Name dosage X vs. placebo)."
  )
)

tbl <- make_table_09(adae = adae, alt_counts_df = adsl, annotations = annotations)
tbl
}
