% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{make_table_09_gtsummary}
\alias{make_table_09_gtsummary}
\title{Engine-Specific Functions: Table 09}
\usage{
make_table_09_gtsummary(
  df,
  denominator = NULL,
  saffl_var = "SAFFL",
  ser_var = "AESER",
  arm_var = "ARM",
  pref_var = "AEDECOD",
  id_var = "USUBJID",
  soc_var = "AESOC",
  lbl_overall = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{denominator}{(\code{data.frame})\cr alternative dataset (typically ADSL) used only to calculate denominator counts.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{ser_var}{(\code{character})\cr Name of the serious adverse event flag variable which must contain at least
"Y" for "Yes".}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{pref_var}{(\code{character})\cr Name of the preferred term variable from \code{adae} to include in the table.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{soc_var}{(\code{character})\cr Name of the system organ class variable from \code{adae} to include in the table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}
}
\value{
\itemize{
\item \code{make_table_09_gtsummary()} returns a \code{gtsummary} object.
}
}
\description{
The table engine used by each engine-specific function is identified by its suffix.
}
\examples{
adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

tbl_gtsummary <- make_table_09_gtsummary(df = adae, denominator = adsl)
tbl_gtsummary
}
\seealso{
\code{\link[=make_table_09]{make_table_09()}}
}
