% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_03.R
\name{preproc_df_table_03}
\alias{preproc_df_table_03}
\title{Pre-Process Data for Table 3 Creation}
\usage{
preproc_df_table_03(
  df,
  arm_var = "ARM",
  id_var = "USUBJID",
  scrnfl_var = "SCRNFL",
  scrnfailfl_var = "SCRNFAILFL",
  scrnfail_var = "SCRNFRS"
)
}
\description{
Pre-Process Data for Table 3 Creation
}
\keyword{internal}
