% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_05.R
\name{preproc_df_table_05}
\alias{preproc_df_table_05}
\title{Pre-Process Data for Table 5 Creation}
\usage{
preproc_df_table_05(
  df,
  saffl_var = "SAFFL",
  trtsdtm_var = "TRTSDTM",
  trtedtm_var = "TRTEDTM",
  u_trtdur = "days"
)
}
\description{
Pre-Process Data for Table 5 Creation
}
\keyword{internal}
