% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_10.R
\name{preproc_df_table_10}
\alias{preproc_df_table_10}
\title{Pre-Process Data for Table 10 Creation}
\usage{
preproc_df_table_10(
  df,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  fmqsc_var = "FMQ01SC",
  fmqnam_var = "FMQ01NAM",
  fmq_scope = "NARROW",
  na_level = "<Missing>"
)
}
\description{
Pre-Process Data for Table 10 Creation
}
\keyword{internal}
