% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_11.R
\name{preproc_df_table_11}
\alias{preproc_df_table_11}
\title{Pre-Process Data for Table 11 Creation}
\usage{
preproc_df_table_11(
  df,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  fmqsc_var = "FMQ01SC",
  fmqnam_var = "FMQ01NAM",
  fmq_scope = "NARROW",
  na_level = "<Missing>"
)
}
\description{
Pre-Process Data for Table 11 Creation
}
\keyword{internal}
