% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_12.R
\name{preproc_df_table_12}
\alias{preproc_df_table_12}
\title{Pre-Process Data for Table 12 Creation}
\usage{
preproc_df_table_12(
  df,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  pref_var = "AEDECOD"
)
}
\description{
Pre-Process Data for Table 12 Creation
}
\keyword{internal}
