% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_32.R, R/fda-table_32_OLD.R
\name{preproc_df_table_32}
\alias{preproc_df_table_32}
\title{Pre-Process Data for Table 32 Creation}
\usage{
preproc_df_table_32(
  df,
  denominator = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  subset = "PARAMCD == 'DIABP' & AVISITN >= 1"
)

preproc_df_table_32(
  df,
  denominator = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  subset = "PARAMCD == 'DIABP' & AVISITN >= 1"
)
}
\arguments{
\item{subset}{(\code{string})\cr selection of both PARAMCD and definition of baseline.}
}
\description{
Pre-Process Data for Table 32 Creation

Pre-Process Data for Table 32 Creation
}
\keyword{internal}
