% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_35.R
\name{preproc_df_table_35}
\alias{preproc_df_table_35}
\title{Pre-Process Data for Table 35 Creation}
\usage{
preproc_df_table_35(
  df,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  soc_var = "AEBODSYS",
  na_level = "<Missing>"
)
}
\description{
Pre-Process Data for Table 35 Creation
}
\keyword{internal}
