% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{renaming_function}
\alias{renaming_function}
\title{Helper function to rename the combined columns}
\usage{
renaming_function(x)
}
\arguments{
\item{x}{character vector containing all column names which start with "combined"}
}
\description{
Helper function to rename the combined columns
}
