\name{SummaryDataFrame-class}
\docType{class}

\alias{class:SummaryDataFrame}
\alias{SummaryDataFrame}
\alias{SummaryDataFrame-class}

\alias{show,SummaryDataFrame-method}
\alias{summary,SummaryDataFrame-method}

\title{SummaryDataFrame: object summaries}

\description{
	An \code{SummaryDataFrame} is a simple extension of the \code{\linkS4class{DataFrame}} class as defined in the 'S4Vectors' package, modified to print more simply, and with additional summary information. It is intended to be appropriate for printing summaries of model objects in a way such that summary statistics can be easily extracted from them later.
}

\author{Kylie A. Bemis}

\seealso{
	\code{\link{DataFrame}}
}

\keyword{classes}
