% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_32.R
\name{tbl_make_table_32}
\alias{tbl_make_table_32}
\alias{make_table_32_gtsummary}
\alias{make_table_32_rtables}
\title{Engine-Specific Functions: Table 32}
\usage{
make_table_32_gtsummary(
  df,
  ard,
  denominator = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  lbl_overall = NULL
)

make_table_32_rtables(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  lbl_overall = NULL,
  risk_diff = NULL,
  prune_0 = FALSE,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr Whether column counts should be printed. Boolean.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add one or more risk difference columns to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the names of the arms to compare to the reference arm. A new column will be added for each
element of \code{arm_y}.
\item \code{col_label}: (optional) labels to use for the risk difference columns. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}. For more than one risk difference column, \code{"arm x vs. arm y"} text will also
be included in the column labels by default. The length of \code{col_label} must be equal to the length of \code{arm_y}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}

\item{prune_0}{(\code{flag})\cr Whether all-zero rows should be removed from the table. Boolean.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\value{
\itemize{
\item \code{make_table_32_gtsummary()} returns a \code{gtsummary} object.
\item \code{make_table_32_rtables()} returns an \code{rtable} object.
}
}
\description{
The table engine used by each engine-specific function is identified by its suffix.
}
\details{
\itemize{
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table "Patients Treated" section are the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are not removed by default (see \code{prune_0} argument).
\item Records with missing treatment start and/or end datetime are excluded from all calculations.
}
}
\examples{
adsl <- random.cdisc.data::cadsl
advs <- random.cdisc.data::cadvs

# gtsummary table --------------
ard <- cardinal:::ard_table_32(df = advs, denominator = adsl)
tbl_gtsummary <- cardinal::make_table_32_gtsummary(
  df = advs,
  ard = ard,
  denominator = adsl
)
tbl_gtsummary

# rtables table ----------------
tbl_rtables <- cardinal::make_table_32_rtables(df = advs, alt_counts_df = adsl)
tbl_rtables

}
\seealso{
\code{\link[=make_table_32]{make_table_32()}}
}
