% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alt_counts_df_preproc}
\alias{alt_counts_df_preproc}
\title{Pre-Process \code{alt_counts_df} for Safety Population}
\usage{
alt_counts_df_preproc(
  alt_counts_df,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = NULL
)
}
\arguments{
\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr Name of the safety flag variable which must contain at least "Y" for "Yes".}
}
\value{
A \code{data.frame} (modified \code{alt_counts_df}) or \code{NULL}.
}
\description{
If \code{alt_counts_df} is not \code{NULL}, will check for required variables (\code{arm_var}) and filter
to include only safety population (\code{saffl_var}, if specified), then apply \code{\link[tern:df_explicit_na]{tern::df_explicit_na()}}.
}
\examples{
adsl <- random.cdisc.data::cadsl
alt_counts_df_preproc(adsl)

}
