% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_convention.R
\name{argument_convention}
\alias{argument_convention}
\title{Standard Arguments}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{adex}{(\code{data.frame})\cr dataset (typically ADEX) required to build table.}

\item{advs}{(\code{data.frame})\cr dataset (typically ADVS) required to build table.}

\item{adsl}{(\code{data.frame})\cr dataset (typically ADSL) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{eosdy_var}{(\code{character})\cr variable denoting last recorded (relative) study day.}

\item{fmqsc_var}{(\code{character})\cr FMQ scope variable to use in table.}

\item{fmqnam_var}{(\code{character})\cr FMQ reference name variable to use in table.}

\item{fmq_scope}{(\code{character})\cr FMQ scope ("NARROW" or "BROAD") to output in table.}

\item{ggtheme}{(\code{theme})\cr a graphical theme as provided by \code{ggplot2} to control styling of the \code{ggplot} object.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{lbl_pref_var}{(\code{character})\cr label corresponding to preferred term variable \code{pref_var} to print in the table.}

\item{lbl_soc_var}{(\code{character})\cr label corresponding to system organ class variable \code{soc_var} to print in the
table.}

\item{lbl_vars}{(\code{vector} of \code{character})\cr labels corresponding to variables in \code{vars} to print
in the table. Labels should be ordered according to the order of variables in \code{vars}.}

\item{na_level}{(\code{character})\cr string to represent missing values.}

\item{na_rm}{(\code{flag})\cr whether \code{NA} levels should be removed from the table.}

\item{pref_var}{(\code{character})\cr preferred term variable from \code{adae} to include in the table.}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add one or more risk difference columns to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the names of the arms to compare to the reference arm. A new column will be added for each
element of \code{arm_y}.
\item \code{col_label}: (optional) labels to use for the risk difference columns. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}. For more than one risk difference column, \code{"arm x vs. arm y"} text will also
be included in the column labels by default. The length of \code{col_label} must be equal to the length of \code{arm_y}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{sex_scope}{(\code{character})\cr Level of \code{SEX} to output in table.}

\item{soc_var}{(\code{character})\cr flag variable used to indicate system organ class.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{trtsdtm_var}{(\code{character})\cr treatment start datetime variable.}

\item{trtedtm_var}{(\code{character})\cr treatment end datetime variable.}

\item{u_trtdur}{(\code{character})\cr unit for duration of treatment. Options are \code{"days"}, \code{"weeks"}, \code{"months"},
and \code{"years"}.}

\item{vars}{(\code{vector} of \code{character})\cr variables from \code{df} to include in the table.}

\item{xticks}{(\code{vector} of \code{numeric})\cr x-axis tick positions. If \code{NA} (default), tick mark positions are
automatically calculated.}

\item{x_lab}{(\code{character})\cr x-axis label.}

\item{y_lab}{(\code{character})\cr y-axis label.}
}
\description{
This documentation lists all the arguments in \code{falcon} that are used repeatedly by functions to output tables.
}
\keyword{internal}
