% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal.R
\docType{package}
\name{cardinal-package}
\alias{cardinal}
\alias{cardinal-package}
\title{cardinal Package}
\description{
Implementation of FDA Safety Tables and Figures
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmaverse/cardinal/}
  \item Report bugs at \url{https://github.com/pharmaverse/cardinal/issues}
}

}
\author{
\strong{Maintainer}: Pawel Rucki \email{pawel.rucki@roche.com}

}
\keyword{internal}
