% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{count_subjects}
\alias{count_subjects}
\title{Helper function for \code{create_table_09_data()}
Used for counting subjects per group}
\usage{
count_subjects(
  adae,
  arm_var,
  id_var,
  sub_level_vars = NULL,
  lbl_overall = NULL,
  risk_diff = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{sub_level_vars}{(\code{NULL} or \code{character}) specifying the sub group for counted subjects}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add one or more risk difference columns to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the names of the arms to compare to the reference arm. A new column will be added for each
element of \code{arm_y}.
\item \code{col_label}: (optional) labels to use for the risk difference columns. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}. For more than one risk difference column, \code{"arm x vs. arm y"} text will also
be included in the column labels by default. The length of \code{col_label} must be equal to the length of \code{arm_y}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}
}
\value{
A \code{data.frame} containing the number of subjects with \code{sub_level_vars} events
}
\description{
Helper function for \code{create_table_09_data()}
Used for counting subjects per group
}
