% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{create_table_09_data}
\alias{create_table_09_data}
\title{Helper function to create the data for \code{make_table08_gt()}}
\usage{
create_table_09_data(
  adae,
  alt_counts_df,
  arm_var,
  id_var,
  soc_var,
  pref_var,
  lbl_overall = NULL,
  risk_diff = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{pref_var}{(\code{character})\cr preferred term variable from \code{adae} to include in the table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}
}
\value{
list containing the counted data to be displayed for table 9 and
a \code{data.frame} containing information about the total N for each group
}
\description{
Helper function to create the data for \code{make_table08_gt()}
}
\details{
If \code{lbl_overall} is not \code{NULL} only the data for the overall column will be generated
}
