\name{defunct}
\alias{Cardinal-defunct}

\alias{class:Binmat}
\alias{Binmat}
\alias{Binmat-class}

\alias{topLabels}
\alias{topLabels,ANY-method}

\title{Defunct functions and methods in Cardinal}

\description{
  These functions are defunct and are no longer available.

  \describe{
        \item{\code{Binmat}:}{\code{\link{matter_mat}}}
        \item{\code{topLabels}:}{\code{\link{topFeatures}}}
    }
}
