\name{defunct}
\alias{Cardinal-defunct}

\alias{standardizeSamples}
\alias{select,SImageSet-method}

\title{Defunct functions and methods in Cardinal}

\description{
  These functions are defunct and are no longer available.

  \describe{
        \item{\code{standardizeSamples}:}{\code{\link{standardizeRuns}}}
        \item{\code{select}:}{\code{\link{selectROI}} (for 'SImageSet')}
    }
}
