\name{deprecated}
\alias{deprecated}

% legacy classes

\alias{class:Hashmat}
\alias{Hashmat}
\alias{Hashmat-class}

\alias{class:IAnnotatedDataFrame}
\alias{IAnnotatedDataFrame}
\alias{IAnnotatedDataFrame-class}

\alias{class:ImageData}
\alias{ImageData}
\alias{ImageData-class}

\alias{class:iSet}
\alias{iSet}
\alias{iSet-class}

\alias{class:MIAPE-Imaging}
\alias{MIAPE-Imaging}
\alias{MIAPE-Imaging-class}

\alias{class:MSImageData}
\alias{MSImageData}
\alias{MSImageData-class}

\alias{class:MSImageProcess}
\alias{MSImageProcess}
\alias{MSImageProcess-class}

\alias{class:MSImageSet}
\alias{MSImageSet}
\alias{MSImageSet-class}

\alias{class:ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}

\alias{class:SImageData}
\alias{SImageData}
\alias{SImageData-class}

\alias{class:SImageSet}
\alias{SImageSet}
\alias{SImageSet-class}

% defunct

\alias{simulateSpectrum}
\alias{aggregate,SpectralImagingExperiment-method}
\alias{featureApply}
\alias{featureApply,ANY-method}
\alias{iData}
\alias{iData<-}
\alias{iData,ANY-method}
\alias{iData<-,ANY-method}
\alias{mzAlign}
\alias{mzAlign,ANY-method}
\alias{mzAlign,ANY-method}
\alias{mzBin}
\alias{mzBin,ANY-method}
\alias{mzBin,ANY-method}
\alias{mzFilter}
\alias{mzFilter,ANY-method}
\alias{peakBin}
\alias{peakBin,ANY-method}
\alias{peakFilter}
\alias{peakFilter,ANY-method}
\alias{pixelApply}
\alias{pixelApply,ANY-method}
\alias{spatialApply}
\alias{spatialApply,ANY-method}

\alias{slice}

\alias{setCardinalNumBlocks}
\alias{getCardinalNumBlocks}

\title{Deprecated and defunct objects in Cardinal}

\description{
	These functions are provided for compatibility with older versions of Cardinal, and will be removed in the future.
}
