\name{deprecated}
\alias{Cardinal-deprecated}

\alias{generateImage}
\alias{generateSpectrum}

\title{Deprecated functions and methods in Cardinal}

\description{
  These functions are provided for compatibility with older versions of Cardinal, and will be defunct at the next release.

  \describe{
        \item{\code{generateSpectrum}:}{\code{\link{simulateSpectrum}}}
        \item{\code{generateImage}:}{\code{\link{simulateImage}}}
    }
}
