\name{deprecated}
\alias{Cardinal-deprecated}

\alias{standardizeSamples}
\alias{select,SImageSet-method}

\title{Deprecated functions and methods in Cardinal}

\description{
  These functions are provided for compatibility with older versions of Cardinal, and will be defunct at the next release.

  \describe{
        \item{\code{standardizeSamples}:}{\code{\link{standardizeRuns}}}
        \item{\code{select}:}{\code{\link{selectROI}} (for 'SImageSet')}
    }
}
