\name{deprecated}
\alias{Deprecated}

\alias{class:AnnotatedImage}
\alias{AnnotatedImage}
\alias{AnnotatedImage-class}

\alias{length,AnnotatedImage-method}
\alias{coord,AnnotatedImage-method}
\alias{coord<-,AnnotatedImage-method}
\alias{coordinates,AnnotatedImage-method}
\alias{coordinates<-,AnnotatedImage-method}
\alias{resolution,AnnotatedImage-method}
\alias{resolution<-,AnnotatedImage-method}
\alias{height}
\alias{height<-}
\alias{height,AnnotatedImage-method}
\alias{height<-,AnnotatedImage-method}
\alias{width,AnnotatedImage-method}
\alias{width<-,AnnotatedImage-method}
\alias{range,AnnotatedImage-method}

\alias{class:AnnotatedImagingExperiment}
\alias{AnnotatedImagingExperiment}
\alias{AnnotatedImagingExperiment-class}

\alias{class:AnnotatedImageList}
\alias{AnnotatedImageList}
\alias{AnnotatedImageList-class}

\alias{[,AnnotatedImageList,ANY,ANY,ANY-method}
\alias{cbind,AnnotatedImageList-method}
\alias{rbind,AnnotatedImageList-method}
\alias{combine,AnnotatedImageList,ANY-method}

\alias{resolution,AnnotatedImagingExperiment-method}
\alias{coord,AnnotatedImagingExperiment-method}
\alias{coordinates,AnnotatedImagingExperiment-method}
\alias{width,AnnotatedImagingExperiment-method}
\alias{height,AnnotatedImagingExperiment-method}

\alias{dim,AnnotatedImageList-method}
\alias{dims,AnnotatedImageList-method}
\alias{dims,AnnotatedImagingExperiment-method}

\alias{show,AnnotatedImageList-method}
\alias{show,AnnotatedImagingExperiment-method}

\alias{spatialApply}
\alias{spatialApply-methods}
\alias{spatialApply,SparseImagingExperiment-method}

\title{Deprecated functions and methods in Cardinal}

\description{
	These functions are provided for compatibility with older versions of Cardinal, and will be defunct at the next release.
}
