\name{deprecated}
\alias{Cardinal-deprecated}

\alias{class:Binmat}
\alias{Binmat}
\alias{Binmat-class}

\title{Deprecated functions and methods in Cardinal}

\description{
  These functions are provided for compatibility with older versions of Cardinal, and will be defunct at the next release.

  \describe{
        \item{\code{Binmat}:}{\code{\link{matter_mat}}}
    }
}
