\name{deprecated}
\alias{deprecated}

% legacy classes

\alias{class:Hashmat}
\alias{Hashmat}
\alias{Hashmat-class}

\alias{class:IAnnotatedDataFrame}
\alias{IAnnotatedDataFrame}
\alias{IAnnotatedDataFrame-class}

\alias{class:ImageData}
\alias{ImageData}
\alias{ImageData-class}

\alias{class:iSet}
\alias{iSet}
\alias{iSet-class}

\alias{class:MIAPE-Imaging}
\alias{MIAPE-Imaging}
\alias{MIAPE-Imaging-class}

\alias{class:MSImageData}
\alias{MSImageData}
\alias{MSImageData-class}

\alias{class:MSImageProcess}
\alias{MSImageProcess}
\alias{MSImageProcess-class}

\alias{class:MSImageSet}
\alias{MSImageSet}
\alias{MSImageSet-class}

\alias{class:ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}

\alias{class:SImageData}
\alias{SImageData}
\alias{SImageData-class}

\alias{class:SImageSet}
\alias{SImageSet}
\alias{SImageSet-class}

% deprecated and defunct

\alias{alpha.colors}
\alias{bw.colors}
\alias{col.map}
\alias{color.map}
\alias{darkmode}
\alias{discrete.colors}
\alias{divergent.colors}
\alias{gradient.colors}
\alias{intensity.colors}
\alias{jet.colors}
\alias{lightmode}
\alias{risk.colors}

\alias{baselineReduction}
\alias{baselineReduction<-}
\alias{cvApply}
\alias{gridded}
\alias{gridded<-}
\alias{height}
\alias{height<-}
\alias{instrumentVendor}
\alias{intensityData}
\alias{intensityData<-}
\alias{ionizationType}
\alias{is3D}
\alias{lineScanDirection}
\alias{massAnalyzerType}
\alias{matrixApplication}
\alias{msiInfo}
\alias{mzData}
\alias{mzData<-}
\alias{normalization}
\alias{normalization<-}
\alias{peakBin}
\alias{peakData}
\alias{peakData<-}
\alias{peakPicking}
\alias{peakPicking<-}
\alias{pixelSize}
\alias{resolution}
\alias{resolution<-}
\alias{resultData}
\alias{resultData<-}
\alias{resultNames}
\alias{resultNames<-}
\alias{scanDirection}
\alias{scanPattern}
\alias{scanPolarity}
\alias{scanType}
\alias{smoothing}
\alias{smoothing<-}
\alias{spectrumRepresentation}
\alias{spectrumRepresentation<-}

\alias{simulateSpectrum}

\alias{aggregate,SpectralImagingExperiment-method}
\alias{featureApply}
\alias{featureApply,SpectralImagingExperiment-method}
\alias{iData}
\alias{iData<-}
\alias{iData,ANY-method}
\alias{iData<-,ANY-method}
\alias{mzAlign}
\alias{mzAlign,MSImagingExperiment,missing-method}
\alias{mzAlign,MSImagingExperiment,numeric-method}
\alias{mzBin}
\alias{mzBin,MSImagingExperiment,missing-method}
\alias{mzBin,MSImagingExperiment,numeric-method}
\alias{mzFilter}
\alias{mzFilter,MSImagingExperiment-method}
\alias{peakFilter}
\alias{peakFilter,MSImagingExperiment-method}
\alias{pixelApply}
\alias{pixelApply,SpectralImagingExperiment-method}
\alias{smoothSignal}
\alias{smoothSignal,SpectralImagingExperiment-method}
\alias{spatialApply}
\alias{spatialApply,SpectralImagingExperiment-method}

\alias{setCardinalNumBlocks}
\alias{getCardinalNumBlocks}

\title{Deprecated and defunct objects in Cardinal}

\description{
	These functions are provided for compatibility with older versions of Cardinal, and will be removed in the future.
}
