\name{legacy}
\alias{Cardinal-legacy}

\alias{batchProcess}
\alias{batchProcess-methods}
\alias{batchProcess,MSImageSet-method}

\alias{class:Hashmat}
\alias{Hashmat}
\alias{Hashmat-class}

\alias{class:IAnnotatedDataFrame}
\alias{IAnnotatedDataFrame}
\alias{IAnnotatedDataFrame-class}

\alias{class:ImageData}
\alias{ImageData}
\alias{ImageData-class}

\alias{class:iSet}
\alias{iSet}
\alias{iSet-class}

\alias{class:MIAPE-Imaging}
\alias{MIAPE-Imaging}
\alias{MIAPE-Imaging-class}

\alias{class:MSImageData}
\alias{MSImageData}
\alias{MSImageData-class}

\alias{class:MSImageProcess}
\alias{MSImageProcess}
\alias{MSImageProcess-class}

\alias{class:MSImageSet}
\alias{MSImageSet}
\alias{MSImageSet-class}

\alias{class:ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}

\alias{class:SImageData}
\alias{SImageData}
\alias{SImageData-class}

\alias{class:SImageSet}
\alias{SImageSet}
\alias{SImageSet-class}

\title{Legacy classes and methods in Cardinal}

\description{
  These classes and methods are deprecated and should no longer be used.

  The class definitions will remain for compatibility with \code{CardinalWorkflows} datasets and other older serialized datasets.

  Objects of these classes should be updated to a supported version of the class.

  For example, for an object of class \code{MSImageSet}:

  \code{object <- as(object, "MSImagingExperiment")}
}
