% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-fig_02.R
\name{make_fig_02}
\alias{make_fig_02}
\title{FDA Figure 2: Time to Last Follow Up, Safety Population, Pooled Analyses}
\usage{
make_fig_02(
  df,
  arm_var = "ARM",
  id_var = "USUBJID",
  saffl_var = "SAFFL",
  eosdy_var = "EOSDY",
  u_trtdur = "days",
  x_lab = paste0("Time from first dose (", u_trtdur, ")"),
  y_lab = "Percent of Patients (\%)",
  xticks = NA,
  add_table = TRUE,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{eosdy_var}{(\code{character})\cr variable denoting last recorded (relative) study day.}

\item{u_trtdur}{(\code{character})\cr unit for duration of treatment. Options are \code{"days"}, \code{"weeks"}, \code{"months"},
and \code{"years"}.}

\item{x_lab}{(\code{character})\cr x-axis label.}

\item{y_lab}{(\code{character})\cr y-axis label.}

\item{xticks}{(\code{vector} of \code{numeric})\cr x-axis tick positions. If \code{NA} (default), tick mark positions are
automatically calculated.}

\item{add_table}{(\code{flag})\cr whether "Number of Patients" table should be printed under the plot.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the figure. Valid annotation types
are \code{title}, \code{subtitles}, and \code{caption}. Each name-value pair should use the annotation type as name and the
desired string as value.}
}
\value{
A \code{ggplot2} object.
}
\description{
FDA Figure 2: Time to Last Follow Up, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{df} must contain the variables specified by \code{arm_var}, \code{id_var}, \code{saffl_var}, and \code{eosdy_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item It is assumed that every record for a unique patient in \code{df} has the same last recorded (relative) study day.
\item Values in the "Number of Patients" table are the number of patients for each arm with treatment duration equal to
or greater than the given time (times corresponding to the figure's x-ticks labels).
\item Records with missing treatment start and/or end datetime are excluded from all calculations.
}
}
\examples{
adsl <- random.cdisc.data::cadsl
adsl$EOSDY <- sample(0:400, nrow(adsl), replace = TRUE)

fig <- make_fig_02(df = adsl)
fig

}
