% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-fig_14.R
\name{make_fig_14}
\alias{make_fig_14}
\title{FDA Figure 14: Mean and 95\% Confidence Interval of Systolic Blood Pressure Over Time
by Treatment Arm, Safety Population, Trial X}
\usage{
make_fig_14(
  df,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  visit_var = "AVISIT",
  paramcd_val = "SYSBP",
  add_cond = NULL,
  x_lab = "",
  y_lab = NULL,
  yticks = NA,
  ggtheme = NULL,
  add_table = TRUE,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{visit_var}{(\code{character})\cr visit variable to put on the x-axis}

\item{add_cond}{(\code{expr})\cr expression that provides additional filters for the analysis
(for instance on \code{ATPT} or \code{VSPOS}).}

\item{x_lab}{(\code{character})\cr x-axis label.}

\item{y_lab}{(\code{character})\cr y-axis label.}

\item{yticks}{(\code{vector} of \code{numeric})\cr y-axis tick positions. If \code{NA} (default), tick mark positions are
automatically calculated.}

\item{ggtheme}{(\code{theme})\cr a graphical theme as provided by \code{ggplot2} to control styling of the \code{ggplot} object.}

\item{add_table}{(\code{flag})\cr whether "Mean Value" and "Number of Patients" tables should be printed under the plot.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the figure. Valid annotation types
are \code{title}, \code{subtitles}, and \code{caption}. Each name-value pair should use the annotation type as name and the
desired string as value.}
}
\value{
A \code{ggplot2} object.
}
\description{
FDA Figure 14: Mean and 95\% Confidence Interval of Systolic Blood Pressure Over Time
by Treatment Arm, Safety Population, Trial X
}
\details{
\itemize{
\item \code{df} must contain the variables \code{AVAL} and \code{PARAMCD}, and the variables specified by \code{arm_var},
\code{saffl_var}, \code{visit_var}, and \code{add_cond}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item It is assumed that \code{df} contains one unique record per patient.
}
}
\examples{
advs <- random.cdisc.data::cadvs

fig <- make_fig_14(
  df = advs,
  add_cond = expr("ONTRTFL == 'Y' | ABLFL == 'Y'"),
  add_table = TRUE,
  yticks = c(135, 140, 145, 150, 155, 160)
)
fig

}
