% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_02.R
\name{make_table_02}
\alias{make_table_02}
\alias{make_table_02_rtables}
\title{FDA Table 2: Baseline Demographic and Clinical Characteristics, Safety Population, Pooled Analyses}
\usage{
make_table_02(
  df,
  return_ard = TRUE,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  vars = c("SEX", "AGE", "AGEGR1", "RACE", "ETHNIC", "COUNTRY")
)

make_table_02_rtables(
  df,
  show_colcounts = TRUE,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  vars = c("SEX", "AGE", "AGEGR1", "RACE", "ETHNIC", "COUNTRY"),
  lbl_vars = formatters::var_labels(df, fill = TRUE)[vars],
  lbl_overall = "Total Population",
  na_rm = FALSE,
  prune_0 = TRUE,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr Name of the safety flag variable which must contain at least "Y" for "Yes".}

\item{vars}{(\code{vector} of \code{character})\cr Variables from \code{df} to include in the table.}

\item{show_colcounts}{(\code{flag})\cr Whether column counts should be printed. Boolean.}

\item{lbl_vars}{(\code{vector} of \code{character})\cr Labels corresponding to variables in \code{vars} to print
in the table. Labels should be ordered according to the order of variables in \code{vars}.}

\item{lbl_overall}{(\code{character})\cr Optional. If specified, an overall column will be added to the table with
the given value as the column label.}

\item{na_rm}{(\code{flag})\cr Whether \code{NA} levels should be removed from the table.}

\item{prune_0}{(\code{flag})\cr Whether all-zero rows should be removed from the table. Boolean.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\value{
A \code{gtsummary} table and, if \code{return_ard = TRUE}, an ARD.
If \code{return_ard = TRUE}, they will be returned as a list with named elements \code{table} and \code{ard}.

\itemize{
\item \code{make_table_02_rtables()} returns an \code{rtable} object.
}
}
\description{
The table engine used by each engine-specific function is identified by its suffix.
}
\details{
\itemize{
\item \code{df} must contain the variables specified by \code{arm_var}, \code{saffl_var}, and
variables specified by the \code{vars} argument.
\item \code{vars} must be a character vector specifying the variables to evaluate.
\item \code{return_ard} set to \code{TRUE} or \code{FALSE}; whether the intermediate ARD
object should be returned.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
}

\itemize{
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item All-zero rows are not removed by default (see \code{prune_0} argument).
}
}
\examples{
library(dplyr)

adsl <- random.cdisc.data::cadsl \%>\%
  mutate(AGEGR1 = as.factor(case_when(
    AGE >= 17 & AGE < 65 ~ ">=17 to <65",
    AGE >= 65 ~ ">=65",
    AGE >= 65 & AGE < 75 ~ ">=65 to <75",
    AGE >= 75 ~ ">=75"
  )))

tbl <- make_table_02(df = adsl)
tbl

adsl <- random.cdisc.data::cadsl


# rtables table ----------------
adsl <- random.cdisc.data::cadsl \%>\%
  mutate(AGEGR1 = as.factor(case_when(
    AGE >= 17 & AGE < 65 ~ ">=17 to <65",
    AGE >= 65 ~ ">=65",
    AGE >= 65 & AGE < 75 ~ ">=65 to <75",
    AGE >= 75 ~ ">=75"
  )))
tbl_rtables <- make_table_02_rtables(df = adsl)
tbl_rtables
}
\seealso{
\code{\link{make_table_02}}

\code{\link[=make_table_02]{make_table_02()}}
}
