% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_02.R
\name{make_table_02}
\alias{make_table_02}
\title{FDA Table 2: Baseline Demographic and Clinical Characteristics, Safety Population, Pooled Analyses}
\usage{
make_table_02(
  data,
  show_colcounts = TRUE,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  vars = c("SEX", "AGE", "RACE", "ETHNIC", "COUNTRY"),
  lbl_vars = formatters::var_labels(data, fill = TRUE)[vars],
  lbl_overall = "Total Population",
  na_rm = FALSE,
  ...,
  return_ard = TRUE
)
}
\arguments{
\item{show_colcounts}{(\code{flag})\cr Whether column counts should be printed. Boolean.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{vars}{(\code{vector} of \code{character})\cr Variables from \code{df} to include in the table.}

\item{lbl_vars}{(\code{vector} of \code{character})\cr labels corresponding to variables in \code{vars} to print
in the table. Labels should be ordered according to the order of variables in \code{vars}.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{na_rm}{(\code{flag})\cr whether \code{NA} levels should be removed from the table.}
}
\value{
\itemize{
\item \code{make_table_02} returns a list containing a \code{gtsummary} table object.
The intermediary \code{ARD} object can also be returned with \code{return_ard} set to \code{TRUE}.
}
}
\description{
FDA Table 2: Baseline Demographic and Clinical Characteristics, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{data} must contain the variables specified by \code{vars}.
\item \code{return_ard} set to \code{TRUE} or \code{FALSE}; whether the intermediate ARD object should be returned.
}
}
\section{Functions}{
\itemize{
\item \code{make_table_02()}: Create FDA table 2 using an ARD.

}}
\examples{
library(dplyr)

adsl <- random.cdisc.data::cadsl \%>\%
  mutate(AGEGR1 = as.factor(case_when(
    AGE >= 17 & AGE < 65 ~ ">=17 to <65",
    AGE >= 65 ~ ">=65",
    AGE >= 65 & AGE < 75 ~ ">=65 to <75",
    AGE >= 75 ~ ">=75"
  )))

tbl <- make_table_02(data = adsl)
tbl

}
