% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_03.R
\name{make_table_03}
\alias{make_table_03}
\title{FDA Table 3: Patient Screening and Enrollment, Trials A and B}
\usage{
make_table_03(
  df,
  alt_counts_df = NULL,
  show_colcounts = FALSE,
  arm_var = "ARM",
  lbl_overall = NULL,
  prune_0 = TRUE,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 3: Patient Screening and Enrollment, Trials A and B
}
\details{
\itemize{
\item \code{adsl} must contain \code{USUBJID} and the variables specified by \code{vars} and \code{arm_var}.
\item If specified, \code{alt_counts_df} must contain \code{USUBJID}, and the variable specified by \code{arm_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm.
\item Numbers in table for non-numeric variables represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")

adsl <- adsl \%>\%
  dplyr::mutate(
    SCRNFL = sample(c(TRUE, NA), size = nrow(adsl), replace = TRUE),
    SCRNFLRS = sample(c(
      "Inclusion/exclusion criteria not met",
      "Patient noncompliance",
      "Consent withdrawn",
      "Other"
    ), size = nrow(adsl), replace = TRUE),
    ENRLFL = sample(c(TRUE, NA), size = nrow(adsl), replace = TRUE),
    RANDFL = sample(c(TRUE, NA), size = nrow(adsl), replace = TRUE)
  )

labels <- c(
  "SCRNFL" = "Patients screened",
  "SCRNFLRS" = "Reason for SCRNFL=N",
  "ENRLFL" = "Patients enrolled",
  "RANDFL" = "Patients randomized"
)

formatters::var_labels(adsl)[names(labels)] <- labels

make_table_03(df = adsl)

}
