% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_05.R
\name{make_table_05}
\alias{make_table_05}
\title{FDA Table 5: Duration of Treatment Exposure, Safety Population, Pooled Analyses}
\usage{
make_table_05(
  df,
  alt_counts_df = NULL,
  return_ard = TRUE,
  arm_var = "ARM",
  id_var = "USUBJID",
  saffl_var = "SAFFL",
  trtsdtm_var = "TRTSDTM",
  trtedtm_var = "TRTEDTM",
  u_trtdur = "days",
  ...
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{trtsdtm_var}{(\code{character})\cr treatment start datetime variable.}

\item{trtedtm_var}{(\code{character})\cr treatment end datetime variable.}

\item{u_trtdur}{(\code{character})\cr unit for duration of treatment. Options are \code{"days"}, \code{"weeks"}, \code{"months"},
and \code{"years"}.}

\item{...}{arguments passed to table engine-specific functions. See \code{\link{tbl_make_table_05}} for details.}
}
\value{
A \code{gtsummary} table and, if \code{return_ard = TRUE}, an ARD.
If \code{return_ard = TRUE}, they will be returned as a list with named elements \code{table} and \code{ard}.
}
\description{
FDA Table 5: Duration of Treatment Exposure, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{df} must contain the variables specified by \code{arm_var}, \code{saffl_var}, \code{id_var}, \code{trtsdtm_var},
and \code{trtedtm_var}.
\item If specified, \code{alt_counts_df} must contain the variables specified by \code{arm_var}, \code{id_var} and \code{saffl_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
}
}
\examples{
adsl <- random.cdisc.data::cadsl

tbl <- make_table_05(df = adsl)
tbl

}
\seealso{
\code{\link{tbl_make_table_05}}
}
