% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_06.R
\name{make_table_06}
\alias{make_table_06}
\title{FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses}
\usage{
make_table_06(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  lbl_overall = NULL,
  prune_0 = FALSE,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr data required to build table.}

\item{alt_counts_df}{(\code{character}) alternative (full) data (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotations types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{df} is typically ADAE.
\item \code{alt_counts_df} is typically ADSL.
\item \code{df} must contain the variables \code{SAFFL}, \code{TRTEMFL}, \code{AESEV}, \code{AESER}, \code{AESDTH}, \code{AESLIFE}, \code{AESHOSP}, \code{AESDISAB},
\code{AESCONG}, \code{AESMIE}, \code{AESACN}, and the variable specified by \code{arm_var}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are not removed by default (see \code{prune_0} argument).
}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

tbl <- make_table_06(df = adae, alt_counts_df = adsl)
tbl

}
