% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_06.R
\name{make_table_06}
\alias{make_table_06}
\title{FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses}
\usage{
make_table_06(
  df,
  denominator,
  return_ard = TRUE,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  sae_var = "AESER",
  sae_cat_vars = list(`SAEs with fatal outcome` = "AESDTH", `Life-threatening SAEs` =
    "AESLIFE", `SAEs requiring hospitalization` = "AESHOSP",
    `SAEs resulting in substantial disruption of normal life functions` = "AESDISAB",
    `Congenital anomaly or birth defect` = "AESCONG", Other = "AESMIE"),
  disc_var = "AEACNOTH",
  dose_mod_var = "AEACN",
  dose_mod_cat_labels = list(`DRUG INTERRUPTED` =
    "AE leading to interruption of study drug", `DOSE REDUCED` =
    "AE leading to reduction of study drug", `DOSE RATE REDUCED` =
    "AE leading to dose delay of study drug", `DOSE INCREASED` = "Other"),
  sev_var = "AESEV"
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{denominator}{(\code{data.frame})\cr alternative dataset (typically ADSL) used only to calculate denominator counts.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{id_var}{(\code{character})\cr Name of the unique subject identifiers variable.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr Name of the safety flag variable which must contain at least "Y" for "Yes".}

\item{sae_var}{(\code{character})\cr Name of the flag variable indicating whether an adverse event is serious ("Y") or
not ("N"). Defaults to AESER from an ADAE ADaM dataset.}

\item{sae_cat_vars}{(named \code{list} of \code{character}) \cr List of names of flag variables for categories of serious
adverse events. These variables may contain "Y" or "No". Defaults to AESDTH, AESLIFE, AESHOSP, AESDISAB, AESCONG,
and AESMIE from an ADAE ADaM dataset. Provide labels through names of the list elements.}

\item{disc_var}{(\code{character}) \cr Name of the variable containing information whether a subject discontinued from
the study as result of an adverse event. Defaults to AEACNOTH from an ADAE ADaM dataset.}

\item{dose_mod_var}{(\code{character}) \cr Name of the variable containing information about dose modification for a
subject due to an adverse event. Defaults to AEACN from an ADAE ADaM dataset.}

\item{dose_mod_cat_labels}{(named \code{list} of \code{character}) \cr List of labels for the dose modification categories as
listed in the details section.}

\item{sev_var}{Name of the variable containing information about the severity of an adverse event. Defaults to AESEV
from an ADAE ADaM dataset.}
}
\value{
A \code{gtsummary} table and, if \code{return_ard = TRUE}, an ARD.
If \code{return_ard = TRUE}, objects will be returned as a list with named elements \code{table} and \code{ard}.
}
\description{
FDA Table 6: Overview of Adverse Events, Safety Population, Pooled Analyses
}
\details{
Typically, an ADAE ADaM dataset is passed to \code{df}. \cr
Typically, an ADSL ADaM dataset is passed to \code{denominator}. \cr

The values passed through \code{id_var} and \code{arm_var} must be present in \code{names(denominator)}. \cr
The values passed through \code{saffl_var}, \code{arm_var}, \code{sae_var}, \code{sae_cat_vars}, \code{disc_var}, \code{dose_mod_var}, and
\code{sev_var} must be present in \code{names(df)}. \cr

Discontinuation from the study needs to be indicated by the value "SUBJECT DISCONTINUED FROM STUDY" in the \code{disc_var}
variable. \cr
Dose modification may be "DRUG INTERRUPTED", "DOSE REDUCED", "DOSE RATE REDUCED", "DOSE INCREASED".
}
\examples{
adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

tbl <- make_table_06(df = adae, denominator = adsl)
tbl
}
