% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{make_table_09}
\alias{make_table_09}
\title{FDA Table 9: Patients With Serious Adverse Events by System Organ Class and Preferred Term,
Safety Population, Pooled Analyses}
\usage{
make_table_09(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  pt_var = "AETERM",
  lbl_pt_var = formatters::var_labels(adae, fill = TRUE)[pt_var],
  lbl_overall = NULL,
  prune_0 = TRUE,
  annotations = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr data (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character}) alternative (full) data (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{pt_var}{(\code{character})\cr preferred term variable from \code{df} to include in the table.}

\item{lbl_pt_var}{(\code{character})\cr label corresponding to preferred term variable \code{pt_var} to print in the table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotations types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 9: Patients With Serious Adverse Events by System Organ Class and Preferred Term,
Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{adae} must contain \code{SAFFL}, \code{USUBJID}, \code{AESER}, \code{AESOC}, and the variables specified
by \code{pt_var} and \code{arm_var}.
\item \code{alt_counts_df} is typically ADSL.
\item Columns are split by arm. Overall population column is not included by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

tbl <- make_table_09(adae = adae, alt_counts_df = adsl)
tbl

}
