% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_09.R
\name{make_table_09}
\alias{make_table_09}
\alias{make_table_09_gtsum}
\title{FDA Table 9: Patients With Serious Adverse Events by System Organ Class and Preferred Term,
Safety Population, Pooled Analyses}
\usage{
make_table_09(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  pref_var = "AEDECOD",
  lbl_pref_var = formatters::var_labels(adae, fill = TRUE)[pref_var],
  lbl_overall = NULL,
  risk_diff = NULL,
  prune_0 = TRUE,
  annotations = NULL
)

make_table_09_gtsum(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  saffl_var = "SAFFL",
  ser_var = "AESER",
  arm_var = "ARM",
  pref_var = "AEDECOD",
  id_var = "USUBJID",
  soc_var = "AESOC",
  lbl_pref_var = formatters::var_labels(adae, fill = TRUE)[pref_var],
  lbl_soc_var = formatters::var_labels(adae, fill = TRUE)[soc_var],
  lbl_overall = NULL,
  annotations = NULL,
  risk_diff = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr safety population flag variable from \code{adae} to include in the table.}

\item{pref_var}{(\code{character})\cr preferred term variable from \code{adae} to include in the table.}

\item{lbl_pref_var}{(\code{character})\cr label corresponding to preferred term variable \code{pref_var} to print in the table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{risk_diff}{(\code{list} of \code{character} vectors)\cr List of character vectors. Each vector must be
of length 2 and contain the name of treatment arms to calculate risk difference and its 95\% CI for. Those names
must exist in the \code{arm_var} column of the dataset specified in \code{adae}.}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, and a list of characters called \code{footnotes}.
Each name-value pair should use the annotation type as name and the desired string as the value.}

\item{ser_var}{(\code{character})\cr serious Event variable from \code{adae} to include in the table.}

\item{soc_var}{(\code{character})\cr system organ class variable from \code{adae} to include in the table.}

\item{lbl_soc_var}{(\code{character})\cr label corresponding to system organ class variable
\code{soc_var} to print in the table.}
}
\value{
An \code{rtable} object.

\itemize{
\item \code{make_table_32_gtsum} returns a \code{gt} object
}
}
\description{
FDA Table 9: Patients With Serious Adverse Events by System Organ Class and Preferred Term,
Safety Population, Pooled Analyses

Function to create FDA Table 9: Patients With Serious Adverse Events
by System Organ Class and Preferred Term,
Safety Population, Pooled Analyses with /code{gt}
}
\details{
\itemize{
\item \code{adae} must contain \code{AESER}, \code{AESOC}, and the variables specified by \code{pref_var}, \code{id_var}, \code{arm_var},
and \code{saffl_var}.
\item If specified, \code{alt_counts_df} must contain the variables specified by \code{arm_var}, \code{id_var}, and \code{saffl_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\section{Functions}{
\itemize{
\item \code{make_table_09_gtsum()}: Create FDA table 9 using functions from \code{gt}.

}}
\examples{
adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

tbl <- make_table_09(adae = adae, alt_counts_df = adsl)
tbl

adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

tbl <- make_table_09_gtsum(adae = adae, alt_counts_df = adsl)
tbl

annotations <- list(
  title = "Table 9. Patients With Serious Adverse Events",
  subtitle = "by System Organ Class and Preferred Term, Safety Population, Pooled Analyses",
  footnotes = list(
    "Duration = [e.g., X week double-blind treatment period or median and a range
    indicating pooled trial durations].",
    "Difference is shown between [treatment arms] (e.g., difference is shown
    between Drug Name dosage X vs. placebo)."
  )
)
risk_diff <- list(c("A: Drug X", "C: Combination"), c("A: Drug X", "B: Placebo"))
tbl <- make_table_09_gtsum(
  adae = adae, alt_counts_df = adsl, annotations = annotations,
  risk_diff = risk_diff
)
tbl
}
