% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_10.R
\name{make_table_10}
\alias{make_table_10}
\title{FDA Table 10: Patients With Serious Adverse Events by System Organ Class and
FDA Medical Query (Narrow), Safety Population, Pooled Analyses}
\usage{
make_table_10(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  fmqsc_var = "FMQ01SC",
  fmqnam_var = "FMQ01NAM",
  fmq_scope = "NARROW",
  lbl_overall = NULL,
  prune_0 = TRUE,
  na_level = "<Missing>",
  annotations = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr Adverse events dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr Optional. Alternative dataset (typically ADSL) used only to calculate column
counts.}

\item{show_colcounts}{(\code{flag})\cr Whether column counts should be printed. Boolean.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{fmqsc_var}{(\code{character})\cr FMQ scope variable to use in table.}

\item{fmqnam_var}{(\code{character})\cr FMQ reference name variable to use in table.}

\item{fmq_scope}{(\code{character})\cr FMQ scope ("NARROW" or "BROAD") to output in table.}

\item{lbl_overall}{(\code{character})\cr Optional. If specified, an overall column will be added to the table with
the given value as the column label.}

\item{prune_0}{(\code{flag})\cr Whether all-zero rows should be removed from the table. Boolean.}

\item{na_level}{(\code{character})\cr String to represent missing values.}

\item{annotations}{(named \code{list} of \code{character})\cr List of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 10: Patients With Serious Adverse Events by System Organ Class and
FDA Medical Query (Narrow), Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{adae} must contain the variables \code{SAFFL}, \code{USUBJID}, \code{AEBODSYS}, \code{AESER}, and the variables specified by
\code{arm_var}, \code{fmqsc_var}, and \code{fmqnam_var}.
\item \code{fmqsc_var} must contain "BROAD" or "NARROW" values, one of which will be displayed in the table. Narrow is
selected by default (see \code{fmq_scope} argument).
\item If specified, \code{alt_counts_df} must contain variables \code{SAFFL} and \code{USUBJID}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
library(dplyr)

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

set.seed(1)
adae <- adae \%>\%
  dplyr::rename(FMQ01SC = SMQ01SC) \%>\%
  dplyr::mutate(
    AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE),
    FMQ01NAM = sample(c("FMQ1", "FMQ2", "FMQ3"), size = nrow(adae), replace = TRUE)
  )
adae$FMQ01SC[is.na(adae$FMQ01SC)] <- "NARROW"

tbl <- make_table_10(adae = adae, alt_counts_df = adsl)
tbl

}
