% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_10.R
\name{make_table_10}
\alias{make_table_10}
\alias{make_table_10_gtsum}
\title{FDA Table 10: Patients With Serious Adverse Events by System Organ Class and
FDA Medical Query (Narrow), Safety Population, Pooled Analyses}
\usage{
make_table_10(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  soc_var = "AEBODSYS",
  fmqsc_var = "FMQ01SC",
  fmqnam_var = "FMQ01NAM",
  fmq_scope = "NARROW",
  lbl_overall = NULL,
  risk_diff = NULL,
  prune_0 = TRUE,
  na_level = "<Missing>",
  annotations = NULL
)

make_table_10_gtsum(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  saffl_var = "SAFFL",
  ser_var = "AESER",
  arm_var = "ARM",
  id_var = "USUBJID",
  soc_var = "AEBODSYS",
  fmqsc_var = "FMQ01SC",
  fmqnam_var = "FMQ01NAM",
  fmq_scope = "NARROW",
  lbl_soc_var = formatters::var_labels(adae, fill = TRUE)[soc_var],
  lbl_overall = NULL,
  annotations = NULL,
  risk_diff = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr safety population flag variable from \code{adae} to include in the table.}

\item{soc_var}{(\code{character})\cr system organ class variable from \code{adae} to include in the table.}

\item{fmqsc_var}{(\code{character})\cr FMQ scope variable to use in table.}

\item{fmqnam_var}{(\code{character})\cr FMQ reference name variable to use in table.}

\item{fmq_scope}{(\code{character})\cr FMQ scope ("NARROW" or "BROAD") to output in table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{risk_diff}{(\code{list} of \code{character} vectors)\cr List of character vectors. Each vector must be
of length 2 and contain the name of treatment arms to calculate risk difference and its 95\% CI for. Those names
must exist in the \code{arm_var} column of the dataset specified in \code{adae}.}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{na_level}{(\code{character})\cr string to represent missing values.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, and a list of characters called \code{footnotes}.
Each name-value pair should use the annotation type as name and the desired string as the value.}

\item{ser_var}{(\code{character})\cr serious Event variable from \code{adae} to include in the table.}

\item{lbl_soc_var}{(\code{character})\cr label corresponding to system organ class variable
\code{soc_var} to print in the table.}
}
\value{
An \code{rtable} object.

\itemize{
\item \code{make_table_10_gt} returns a \code{gt} object
}
}
\description{
FDA Table 10: Patients With Serious Adverse Events by System Organ Class and
FDA Medical Query (Narrow), Safety Population, Pooled Analyses

FDA Table 10: Patients With Serious Adverse Events by System Organ Class and
FDA Medical Query (Narrow), Safety Population, Pooled Analyses with /code{gt}
}
\details{
\itemize{
\item \code{adae} must contain the variables \code{AEBODSYS}, \code{AESER}, and the variables specified by
\code{arm_var}, \code{id_var}, \code{saffl_var}, \code{fmqsc_var}, and \code{fmqnam_var}.
\item If specified, \code{alt_counts_df} must contain \code{USUBJID} and the variables specified by \code{arm_var} and \code{saffl_var}.
\item \code{fmqsc_var} must contain "BROAD" or "NARROW" values, one of which will be displayed in the table. Narrow is
selected by default (see \code{fmq_scope} argument).
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\section{Functions}{
\itemize{
\item \code{make_table_10_gtsum()}: Create FDA table 10 using functions from \code{gt}.

}}
\examples{
library(dplyr)

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

set.seed(1)
adae <- adae \%>\%
  rename(FMQ01SC = SMQ01SC) \%>\%
  mutate(
    AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE),
    FMQ01NAM = sample(c("FMQ1", "FMQ2", "FMQ3"), size = nrow(adae), replace = TRUE)
  )
adae$FMQ01SC[is.na(adae$FMQ01SC)] <- "NARROW"

tbl <- make_table_10(adae = adae, alt_counts_df = adsl)
tbl

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

set.seed(1)
adae <- adae \%>\%
  rename(FMQ01SC = SMQ01SC) \%>\%
  mutate(
    AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE),
    FMQ01NAM = sample(c("FMQ1", "FMQ2", "FMQ3"), size = nrow(adae), replace = TRUE)
  )
adae$FMQ01SC[is.na(adae$FMQ01SC)] <- "NARROW"

tbl <- make_table_10_gtsum(adae = adae, alt_counts_df = adsl)
tbl

annotations <- list(
  title = "Table 10. Patients With Serious Adverse Events ",
  subtitle = "by System Organ Class and FDA Medical Query (Narrow),
  Safety Population, Pooled Analyses",
  footnotes = list(
    "Duration = [e.g., X week double-blind treatment period or median and a range
    indicating pooled trial durations].",
    "Difference is shown between [treatment arms] (e.g., difference is shown
    between Drug Name dosage X vs. placebo)."
  )
)
risk_diff <- list(c("A: Drug X", "C: Combination"), c("A: Drug X", "B: Placebo"))
tbl <- make_table_10_gtsum(adae,
  alt_counts_df = adsl,
  annotations = annotations,
  risk_diff = risk_diff
)
tbl
}
