% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_11.R
\name{make_table_11}
\alias{make_table_11}
\title{FDA Table 11: Patients with FDA Medical Query (Narrow) Leading to Treatment
Discontinuation, Safety Population, Pooled Analyses}
\usage{
make_table_11(
  adae,
  alt_counts_df = adsl,
  show_colcounts = TRUE,
  arm_var = "ARM",
  fmqsc_var = "FMQ02SC",
  fmqnam_var = "FMQ02NAM",
  lbl_overall = NULL,
  prune_0 = TRUE,
  na_level = "<Missing>",
  annotations = NULL,
  FMQ = "Narrow"
)
}
\description{
FDA Table 11: Patients with FDA Medical Query (Narrow) Leading to Treatment
Discontinuation, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{adae} must contain the variables \code{SAFFL}, \code{USUBJID}, \code{AEBODSYS}, \code{AEACN}
and the variables specified by \code{arm_var}, \code{fmqsc_var}, and \code{fmqnam_var}.
\item the function output FMQ with Treatment Discontinuation (\code{AEACN} = "DRUG WITHDRAWN").
\item \code{fmqsc_var} must contain either "BROAD" or "NARROW" values. By default, the
function outputs "NARROW" for the FMQ (see \code{FMQ} argument).
\item If specified, \code{alt_counts_df} must contain variables \code{SAFFL} and \code{USUBJID}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true)
and \code{"N"} (false). Missing values in flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default
(see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
library(dplyr)
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")
adae <- adae \%>\%
  dplyr::rename(FMQ02SC = SMQ02SC, FMQ02NAM = CQ01NAM) \%>\%
  mutate(AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE))
levels(adae$FMQ02SC) <- c("BROAD", "NARROW")
adae$FMQ02SC[!is.na(adae$FMQ02NAM)] <- "NARROW"
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")

tbl <- make_table_11(
  adae = adae, alt_counts_df = adsl, fmqsc_var = "FMQ02SC",
  fmqnam_var = "FMQ02NAM"
)

tbl
}
