% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_12.R
\name{make_table_12}
\alias{make_table_12}
\title{FDA Table 12: Patients With Adverse Events Leading to Treatment Discontinuation
by System Organ Class and Preferred Term, Safety Population, Pooled Analyses}
\usage{
make_table_12(
  adae,
  alt_counts_df = adsl,
  show_colcounts = TRUE,
  arm_var = "ARM",
  pref_var = "AEDECOD",
  lbl_overall = NULL,
  prune_0 = TRUE,
  annotations = NULL
)
}
\description{
FDA Table 12: Patients With Adverse Events Leading to Treatment Discontinuation
by System Organ Class and Preferred Term, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{adae} must contain \code{SAFFL}, \code{USUBJID}, \code{AEACN}, \code{AESOC}, and the variables specified
by \code{pref_var} and \code{arm_var}.
\item If specified, \code{alt_counts_df} must contain \code{SAFFL}, \code{USUBJID}, and the variable specified by \code{arm_var}.
\item the function output \code{pref_var} with Treatment Discontinuation (\code{AEACN} = "DRUG WITHDRAWN").
\item By default, \code{pref_var} is AEDECOD.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
library(dplyr)
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")

tbl <- make_table_12(adae = adae, alt_counts_df = adsl)
tbl

}
