% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_12.R
\name{make_table_12}
\alias{make_table_12}
\title{FDA Table 12: Patients With Adverse Events Leading to Treatment Discontinuation
by System Organ Class and Preferred Term, Safety Population, Pooled Analyses}
\usage{
make_table_12(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  pref_var = "AEDECOD",
  lbl_overall = NULL,
  prune_0 = TRUE,
  annotations = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr Adverse events dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr Optional. Alternative dataset (typically ADSL) used only to calculate column
counts.}

\item{show_colcounts}{(\code{flag})\cr Whether column counts should be printed. Boolean.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{pref_var}{(\code{character})\cr Name of the preferred term variable from \code{adae} to include in the table.}

\item{lbl_overall}{(\code{character})\cr Optional. If specified, an overall column will be added to the table with
the given value as the column label.}

\item{prune_0}{(\code{flag})\cr Whether all-zero rows should be removed from the table. Boolean.}

\item{annotations}{(named \code{list} of \code{character})\cr List of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 12: Patients With Adverse Events Leading to Treatment Discontinuation
by System Organ Class and Preferred Term, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{adae} must contain \code{SAFFL}, \code{USUBJID}, \code{DCSREAS}, \code{AESOC}, and the variables specified by \code{pref_var}
and \code{arm_var}.
\item If specified, \code{alt_counts_df} must contain \code{SAFFL}, \code{USUBJID}, and the variable specified by \code{arm_var}.
\item A patient is defined as having at least one adverse event leading to treatment discontinuation when they have at
least one record with \code{DCSREAS} = "ADVERSE EVENT".
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")
adae$DCSREAS[is.na(adae$DCSREAS)] <- "ADVERSE EVENT"

tbl <- make_table_12(adae = adae, alt_counts_df = adsl)
tbl

}
