% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_12.R
\name{make_table_12}
\alias{make_table_12}
\title{FDA Table 12: Patients With Adverse Events Leading to Treatment Discontinuation
by System Organ Class and Preferred Term, Safety Population, Pooled Analyses}
\usage{
make_table_12(
  df,
  denominator = NULL,
  return_ard = TRUE,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  pref_var = "AEDECOD"
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{denominator}{(\code{data.frame})\cr alternative dataset (typically ADSL) used only to calculate denominator counts.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{id_var}{(\code{character})\cr Name of the unique subject identifiers variable.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr Name of the safety flag variable which must contain at least "Y" for "Yes".}

\item{pref_var}{(\code{character})\cr Name of the preferred term variable from \code{adae} to include in the table.}
}
\value{
A \code{gtsummary} table and, if \code{return_ard = TRUE}, an ARD.
If \code{return_ard = TRUE}, they will be returned as a list with named elements \code{table} and \code{ard}.
}
\description{
FDA Table 12: Patients With Adverse Events Leading to Treatment Discontinuation
by System Organ Class and Preferred Term, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{df} must contain \code{DCSREAS}, \code{AESOC}, and the variables specified by \code{arm_var}, \code{id_var}, \code{saffl_var},
and \code{pref_var}.
\item If specified, \code{denominator} must contain \code{USUBJID} and the variables specified by \code{arm_var} and \code{saffl_var}.
\item A patient is defined as having at least one adverse event leading to treatment discontinuation when they have at
least one record with \code{DCSREAS = "ADVERSE EVENT"}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
}
}
\examples{
adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae
adae$DCSREAS[is.na(adae$DCSREAS)] <- "ADVERSE EVENT"

tbl <- make_table_12(df = adae, denominator = adsl)
tbl

}
\seealso{
\code{\link{tbl_make_table_12}}
}
