% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_13.R
\name{make_table_13}
\alias{make_table_13}
\title{FDA Table 13: Patients With Common Adverse Events Occurring at >=XX\% Frequency, Safety Population, Pooled Analyses}
\usage{
make_table_13(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  min_freq = 0.05,
  arm_var = "ARM",
  pt_var = "AETERM",
  lbl_pt_var = formatters::var_labels(df, fill = TRUE)[pt_var],
  lbl_overall = NULL,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr data required to build table.}

\item{alt_counts_df}{(\code{character}) alternative (full) data (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{min_freq}{(\code{proportion})\cr minimum \% frequency (fraction of \code{N}) required in any column to include each row.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{pt_var}{(\code{character})\cr preferred term variable from \code{df} to include in the table.}

\item{lbl_pt_var}{(\code{character})\cr label corresponding to preferred term variable \code{pt_var} to print in the table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotations types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 13: Patients With Common Adverse Events Occurring at >=XX\% Frequency, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{df} is typically ADAE.
\item \code{alt_counts_df} is typically ADSL.
\item \code{df} must contain \code{SAFFL} and the variables specified by \code{pt_var} and \code{arm_var}.
\item Columns are split by arm. Overall population column is not included by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

tbl <- make_table_13(df = adae, alt_counts_df = adsl)
tbl

}
