% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_13.R
\name{make_table_13}
\alias{make_table_13}
\title{FDA Table 13: Patients With Common Adverse Events Occurring at >=XX\% Frequency, Safety Population, Pooled Analyses}
\usage{
make_table_13(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  min_freq = 0.05,
  arm_var = "ARM",
  pref_var = "AETERM",
  lbl_pref_var = formatters::var_labels(adae, fill = TRUE)[pref_var],
  lbl_overall = NULL,
  annotations = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr data (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character}) alternative (full) dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{min_freq}{(\code{proportion})\cr minimum \% frequency (fraction of \code{N}) required in any column to include each row.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{pref_var}{(\code{character})\cr preferred term variable from \code{adae} to include in the table.}

\item{lbl_pref_var}{(\code{character})\cr label corresponding to preferred term variable \code{pref_var} to print in the table.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotations types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 13: Patients With Common Adverse Events Occurring at >=XX\% Frequency, Safety Population, Pooled Analyses
}
\details{
\itemize{
\item \code{adae} must contain \code{SAFFL}, \code{USUBJID}, and the variables specified by \code{pref_var} and \code{arm_var}.
\item If specified, \code{alt_counts_df} must contain variables \code{SAFFL} and \code{USUBJID}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

tbl <- make_table_13(adae = adae, alt_counts_df = adsl)
tbl

}
