% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_20.R
\name{make_table_20}
\alias{make_table_20}
\title{FDA Table 20: Adverse Events of Special Interest Assessment, Safety Population, Pooled Analysis (or Trial X)}
\usage{
make_table_20(
  adae,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  pref_var = "AEDECOD",
  aesifl_var = "AESIFL",
  aelabfl_var = "AELABFL",
  lbl_overall = NULL,
  risk_diff = NULL,
  prune_0 = TRUE,
  annotations = NULL
)
}
\arguments{
\item{adae}{(\code{data.frame})\cr dataset (typically ADAE) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{pref_var}{(\code{character})\cr preferred term variable from \code{adae} to include in the table.}

\item{aesifl_var}{(\code{character})\cr variable from \code{adae} that indicates adverse events of special interest.}

\item{aelabfl_var}{(\code{character})\cr variable from \code{adae} that indicates a laboratory assessment.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add one or more risk difference columns to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the names of the arms to compare to the reference arm. A new column will be added for each
element of \code{arm_y}.
\item \code{col_label}: (optional) labels to use for the risk difference columns. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}. For more than one risk difference column, \code{"arm x vs. arm y"} text will also
be included in the column labels by default. The length of \code{col_label} must be equal to the length of \code{arm_y}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\value{
An \code{rtable} object.
}
\description{
FDA Table 20: Adverse Events of Special Interest Assessment, Safety Population, Pooled Analysis (or Trial X)
}
\details{
\itemize{
\item \code{adae} must contain \code{AESEV}, \code{AESER}, \code{AESDTH}, \code{EOSSTT}, \code{AEREL}, and the variables specified by
\code{id_var}, \code{pref_var}, \code{aesifl_var}, \code{aelabfl_var}, \code{arm_var}, and \code{saffl_var}.
\item If specified, \code{alt_counts_df} must contain the variables specified by \code{arm_var}, \code{id_var}, and \code{saffl_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")

set.seed(1)
adae$AESIFL <- ifelse(adae$AESOC \%in\% c("cl A", "cl D"), "Y", "N")
adae$AELABFL <- sample(c("Y", "N"), nrow(adae), replace = TRUE)

tbl <- make_table_20(adae = adae, alt_counts_df = adsl)
tbl

}
