% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_21.R
\name{make_table_21}
\alias{make_table_21}
\title{FDA Table 21. Overview of Serious Adverse Events by Demographic Subgroup, Safety Population, Pooled
Analysis (or Trial X)}
\usage{
make_table_21(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  vars = c("SEX", "AGEGR1", "RACE", "ETHNIC"),
  denom = c("N_s", "N_col", "n"),
  lbl_vars = formatters::var_labels(df, fill = TRUE)[vars],
  prune_0 = FALSE,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{vars}{(\code{vector} of \code{character})\cr variables from \code{df} to include in the table.}

\item{denom}{(\code{character})\cr Denominator to use to calculate fractions. Can be \code{"N_s"} (total \code{df_denom}
subgroup/row counts), \code{"N_col"} (total \code{df} column counts), or \code{"n"} (total \code{df} overall patient count). Note that
\code{df} is filtered to only include serious adverse events (\code{ASER == "Y"}).}

\item{lbl_vars}{(\code{vector} of \code{character})\cr labels corresponding to variables in \code{vars} to print
in the table. Labels should be ordered according to the order of variables in \code{vars}.}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 21. Overview of Serious Adverse Events by Demographic Subgroup, Safety Population, Pooled
Analysis (or Trial X)
}
\details{
\itemize{
\item \code{df} must contain \code{SAFFL} and the variables specified by \code{vars} and \code{arm_var}.
\item If specified, \code{alt_counts_df} must contain \code{SAFFL}, \code{USUBJID}, and the variable specified by \code{arm_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm.
\item Information from either ADSUB or ADVS is generally included into \code{df} prior to analysis.
\item Numbers in the table for non-numeric variables represent the absolute number of patients and fraction of \code{n}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
library(dplyr)

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl") \%>\%
  mutate(AGEGR1 = as.factor(case_when(
    AGE >= 17 & AGE < 65 ~ ">=17 to <65",
    AGE >= 65 ~ ">=65",
    AGE >= 65 & AGE < 75 ~ ">=65 to <75",
    AGE >= 75 ~ ">=75"
  )) \%>\% formatters::with_label("Age Group, years")) \%>\%
  formatters::var_relabel(
    AGE = "Age, years"
  )

adae <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adae")
adae$ASER <- adae$AESER

df <- left_join(adsl, adae, by = intersect(names(adsl), names(adae)))

tbl <- make_table_21(df = df, alt_counts_df = adsl)
tbl

}
