% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_22.R
\name{make_table_22}
\alias{make_table_22}
\title{FDA Table 22. Overview of Adverse Events by Demographic Subgroup, Safety Population, Pooled Analysis (or Trial X)}
\usage{
make_table_22(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  vars = c("SEX", "AGEGR1", "RACE", "ETHNIC"),
  denom = c("N_s", "N_col", "n"),
  lbl_overall = NULL,
  lbl_vars = formatters::var_labels(df, fill = TRUE)[vars],
  prune_0 = FALSE,
  annotations = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr Whether column counts should be printed. Boolean.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr Name of the safety flag variable which must contain at least "Y" for "Yes".}

\item{vars}{(\code{vector} of \code{character})\cr Variables from \code{df} to include in the table.}

\item{denom}{(\code{character})\cr Denominator to use to calculate fractions. Can be \code{"N_s"} (total \code{df_denom}
subgroup/row counts), \code{"N_col"} (total \code{df} column counts), or \code{"n"} (total \code{df} overall patient count).
Note that \code{df} is filtered to only include treatment-emergent adverse events (\code{TRTEMFL == "Y"}).}

\item{lbl_overall}{(\code{character})\cr Optional. If specified, an overall column will be added to the table with
the given value as the column label.}

\item{lbl_vars}{(\code{vector} of \code{character})\cr Labels corresponding to variables in \code{vars} to print
in the table. Labels should be ordered according to the order of variables in \code{vars}.}

\item{prune_0}{(\code{flag})\cr Whether all-zero rows should be removed from the table. Boolean.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\value{
An \code{rtable} object.
}
\description{
FDA Table 22. Overview of Adverse Events by Demographic Subgroup, Safety Population, Pooled Analysis (or Trial X)
}
\details{
\itemize{
\item \code{df} must contain the variables specified by \code{vars}, \code{arm_var}, and \code{saffl_var}.
\item If specified, \code{alt_counts_df} must contain \code{USUBJID} and the variables specified by \code{arm_var} and \code{saffl_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm.
\item Information from either ADSUB or ADVS is generally included into \code{df} prior to analysis.
\item Numbers in the table for non-numeric variables represent the absolute number of patients and fraction of \code{n}.
\item All-zero rows are removed by default (see \code{prune_0} argument).
}
}
\examples{
library(dplyr)

adsl <- random.cdisc.data::cadsl \%>\%
  mutate(AGEGR1 = as.factor(case_when(
    AGE >= 17 & AGE < 65 ~ ">=17 to <65",
    AGE >= 65 ~ ">=65",
    AGE >= 65 & AGE < 75 ~ ">=65 to <75",
    AGE >= 75 ~ ">=75"
  )) \%>\% formatters::with_label("Age Group, years")) \%>\%
  formatters::var_relabel(AGE = "Age, years")

adae <- random.cdisc.data::cadae

df <- left_join(adsl, adae, by = intersect(names(adsl), names(adae)))

tbl <- make_table_22(df = df, alt_counts_df = adsl)
tbl

}
