% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_32.R
\name{make_table_32}
\alias{make_table_32}
\title{FDA Table 32: Percentage of Patients With Maximum Diastolic Blood Pressure by Category of Blood Pressure
Postbaseline, Safety Population, Pooled Analysis}
\usage{
make_table_32(
  df,
  alt_counts_df = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  subset,
  lbl_overall = NULL,
  return_ard = TRUE
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{subset}{(\verb{filter condition})\cr selection of both PARAMCD and definition of baseline}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}
}
\value{
A \code{gtsummary} table and, if \code{return_ard = TRUE}, an ARD.
If \code{return_ard = TRUE}, they will be returned as a list with named elements \code{table} and \code{ard}.
}
\description{
FDA Table 32: Percentage of Patients With Maximum Diastolic Blood Pressure by Category of Blood Pressure
Postbaseline, Safety Population, Pooled Analysis
}
\details{
\itemize{
\item \code{df} must contain \code{PARAMCD}, \code{AVAL}, \code{AVALU}, and the variables specified by \code{saffl_var}, \code{arm_var}, and \code{id_var}.
\code{saffl_var} must be a flag variable.
\item \code{return_ard} set to \code{TRUE} or \code{FALSE}; whether the intermediate ARD object should be returned.
}
}
\examples{
adsl <- random.cdisc.data::cadsl
advs <- random.cdisc.data::cadvs

tbl <- make_table_32(
  df = advs,
  alt_counts_df = adsl,
  subset = (PARAMCD == "DIABP" & AVISITN >= 1)
)
tbl

}
\seealso{
\code{\link{tbl_make_table_32}}
}
