% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_32.R
\name{make_table_32}
\alias{make_table_32}
\title{FDA Table 32: Percentage of Patients With Maximum Diastolic Blood Pressure by Category of Blood Pressure
Postbaseline, Safety Population, Pooled Analysis}
\usage{
make_table_32(
  df,
  alt_counts_df = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  lbl_overall = NULL,
  table_engine = c("rtables", "gtsummary"),
  return_ard = TRUE,
  ...
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}
}
\value{
\itemize{
\item \code{make_table_32} returns an object matching the selected \code{table_engine} argument.
The intermediary \code{ARD} object can also be returned with \code{return_ard} set to \code{TRUE}.
}
}
\description{
FDA Table 32: Percentage of Patients With Maximum Diastolic Blood Pressure by Category of Blood Pressure
Postbaseline, Safety Population, Pooled Analysis
}
\details{
\itemize{
\item \code{data} must contain the variables specified by \code{continuous_vars}, and \code{categorical_vars}.
\item \code{table_engine} must be one of \code{gtsummary}, \code{rtables}, \code{tplyr}.
\item \code{return_ard} set to \code{TRUE} or \code{FALSE}; whether the intermediate ARD object should be returned.
}
}
\section{Functions}{
\itemize{
\item \code{make_table_32()}: Create FDA table 32 using an ARD.

}}
\examples{
adsl <- random.cdisc.data::cadsl
advs <- random.cdisc.data::cadvs

tbl <- make_table_32(df = advs, alt_counts_df = adsl)
tbl

}
