% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_33.R
\name{make_table_33}
\alias{make_table_33}
\title{FDA Table 33: Percentage of Patients Meeting Specific Hypotension Levels Postbaseline,
Safety Population, Pooled Analysis}
\usage{
make_table_33(
  advs,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  lbl_overall = NULL,
  risk_diff = NULL,
  prune_0 = FALSE,
  annotations = NULL
)
}
\arguments{
\item{advs}{(\code{data.frame})\cr dataset (typically ADVS) required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{show_colcounts}{(\code{flag})\cr whether column counts should be printed.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add a risk difference column to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the name of the arm to compare to the reference arm.
\item \code{col_label}: (optional) label to use for the risk difference column. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}

\item{prune_0}{(\code{flag})\cr whether all-zero rows should be removed from the table.}

\item{annotations}{(named \code{list} of \code{character})\cr list of annotations to add to the table. Valid
annotation types are \code{title}, \code{subtitles}, \code{main_footer}, and \code{prov_footer}. Each name-value pair should
use the annotation type as name and the desired string as value.}
}
\description{
FDA Table 33: Percentage of Patients Meeting Specific Hypotension Levels Postbaseline,
Safety Population, Pooled Analysis
}
\details{
\itemize{
\item \code{advs} must contain \code{SAFFL}, \code{USUBJID}, \code{AVISITN}, \code{PARAMCD}, \code{AVAL}, \code{AVALU}, and the variable
specified by \code{arm_var}.
\item If specified, \code{alt_counts_df} must contain \code{SAFFL}, \code{USUBJID}, and the variable specified by \code{arm_var}.
\item Flag variables (i.e. \code{XXXFL}) are expected to have two levels: \code{"Y"} (true) and \code{"N"} (false). Missing values in
flag variables are treated as \code{"N"}.
\item Columns are split by arm. Overall population column is excluded by default (see \code{lbl_overall} argument).
\item Numbers in table for non-numeric variables represent the absolute numbers of patients and fraction of \code{N}.
\item All-zero rows are not removed by default (see \code{prune_0} argument).
}
}
\examples{
adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")
advs <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "advs")

tbl <- make_table_33(advs = advs, alt_counts_df = adsl)
tbl

}
