% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_33.R
\name{make_table_33}
\alias{make_table_33}
\title{FDA Table 33: Percentage of Patients Meeting Specific Hypotension Levels Postbaseline,
Safety Population, Pooled Analysis}
\usage{
make_table_33(
  df,
  denominator = NULL,
  return_ard = TRUE,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  lbl_overall = NULL,
  subset = NULL
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{denominator}{(\code{data.frame})\cr alternative dataset (typically ADSL) used only to calculate denominator counts.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{id_var}{(\code{character})\cr Name of the unique subject identifiers variable.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr Name of the safety flag variable which must contain at least "Y" for "Yes".}

\item{lbl_overall}{(\code{character})\cr Optional. If specified, an overall column will be added to the table with
the given value as the column label.}

\item{subset}{(\code{string})\cr selection of both PARAMCD and definition of baseline.}
}
\value{
A \code{gtsummary} table and, if \code{return_ard = TRUE}, an ARD.
If \code{return_ard = TRUE}, they will be returned as a list with named elements \code{table} and \code{ard}.
}
\description{
FDA Table 33: Percentage of Patients Meeting Specific Hypotension Levels Postbaseline,
Safety Population, Pooled Analysis
}
\details{
\itemize{
\item \code{df} must contain \code{PARAMCD}, \code{AVAL}, \code{AVALU}, and the variables specified by \code{saffl_var}, \code{arm_var}, and \code{id_var}.
\code{saffl_var} must be a flag variable.
\item \code{return_ard} set to \code{TRUE} or \code{FALSE}; whether the intermediate ARD object should be returned.
}
}
\examples{
adsl <- random.cdisc.data::cadsl
advs <- random.cdisc.data::cadvs

tbl <- make_table_33(df = advs, denominator = adsl)
tbl

}
\seealso{
\code{\link{tbl_make_table_33}}
}
