% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_33.R, R/fda-table_33_OLD.R
\name{make_table_33}
\alias{make_table_33}
\title{FDA Table 33: Percentage of Patients Meeting Specific Hypotension Levels Postbaseline,
Safety Population, Pooled Analysis}
\usage{
make_table_33(
  df,
  alt_counts_df = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  subset,
  lbl_overall = NULL,
  table_engine = c("rtables", "gtsummary"),
  return_ard = TRUE,
  ...
)

make_table_33(
  df,
  alt_counts_df = NULL,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  subset,
  lbl_overall = NULL,
  table_engine = c("rtables", "gtsummary"),
  return_ard = TRUE,
  ...
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{alt_counts_df}{(\code{character})\cr alternative dataset (typically ADSL) used only to calculate column counts.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{subset}{(\verb{filter condition})\cr selection of both PARAMCD and definition of baseline}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{table_engine}{(\code{string})\cr table engine used to generate the table. If \code{NULL}, no table is returned. If the
given engine is not available for the table, no table is created and a warning is returned.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{...}{arguments passed to table engine-specific functions.}
}
\value{
A \code{gtsummary} table and, if \code{return_ard = TRUE}, an ARD.
If \code{return_ard = TRUE}, they will be returned as a list with named elements \code{table} and \code{ard}.

\itemize{
\item \code{make_table_33} returns an object matching the selected \code{table_engine} argument.
The intermediary \code{ARD} object can also be returned with \code{return_ard} set to \code{TRUE}.
}
}
\description{
FDA Table 33: Percentage of Patients Meeting Specific Hypotension Levels Postbaseline,
Safety Population, Pooled Analysis

FDA Table 33: Percentage of Patients Meeting Specific Hypotension Levels Postbaseline,
Safety Population, Pooled Analysis
}
\details{
\itemize{
\item \code{df} must contain \code{PARAMCD}, \code{AVAL}, \code{AVALU}, and the variables specified by \code{saffl_var}, \code{arm_var}, and \code{id_var}.
\code{saffl_var} must be a flag variable.
\item \code{return_ard} set to \code{TRUE} or \code{FALSE}; whether the intermediate ARD object should be returned.
}

\itemize{
\item \code{df} must contain \code{PARAMCD}, \code{AVAL}, \code{AVALU}, and the variables specified by \code{saffl_var}, \code{arm_var}, and \code{id_var}.
\code{saffl_var} must be a flag variable.
\item \code{table_engine} must be one of \code{gtsummary} or \code{rtables}.
\item \code{return_ard} set to \code{TRUE} or \code{FALSE}; whether the intermediate ARD object should be returned.
}
}
\section{Functions}{
\itemize{
\item \code{make_table_33()}: Create FDA table 33 using an ARD.

}}
\examples{
adsl <- random.cdisc.data::cadsl
advs <- random.cdisc.data::cadvs

tbl <- make_table_33(df = advs, denominator = adsl)
tbl

adsl <- random.cdisc.data::cadsl
advs <- random.cdisc.data::cadvs

tbl <- make_table_33(
  df = advs,
  alt_counts_df = adsl,
  table_engine = "gtsummary",
  subset = (PARAMCD \%in\% c("DIABP", "SYSBP") & AVISITN >= 1)
)
tbl

}
\seealso{
\code{\link{tbl_make_table_33}}
}
