% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda-table_35.R
\name{make_table_35}
\alias{make_table_35}
\title{Table 35. Patients With Adverse Events by System Organ Class,
Safety Population, Pooled Analysis (or Trial X)}
\usage{
make_table_35(
  df,
  denominator = NULL,
  return_ard = TRUE,
  id_var = "USUBJID",
  arm_var = "ARM",
  saffl_var = "SAFFL",
  soc_var = "AEBODSYS",
  lbl_overall = NULL,
  na_level = "<Missing>"
)
}
\arguments{
\item{df}{(\code{data.frame})\cr dataset required to build table.}

\item{denominator}{(\code{data.frame})\cr alternative dataset (typically ADSL) used only to calculate denominator counts.}

\item{return_ard}{(\code{flag})\cr whether an ARD should be returned. Defaults to \code{TRUE}.}

\item{id_var}{(\code{character})\cr variable used as unique subject identifier.}

\item{arm_var}{(\code{character})\cr Name of the treatment arm variable used to split table into columns.}

\item{saffl_var}{(\code{character})\cr flag variable used to indicate inclusion in safety population.}

\item{soc_var}{(\code{character})\cr Name of the variable that contains the SOC to describe.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{na_level}{(\code{character})\cr string to represent missing values.}
}
\value{
A \code{gtsummary} table and, if \code{return_ard = TRUE}, an ARD.
If \code{return_ard = TRUE}, they will be returned as a list with named elements \code{table} and \code{ard}.
}
\description{
Table 35. Patients With Adverse Events by System Organ Class,
Safety Population, Pooled Analysis (or Trial X)
}
\details{
\itemize{
\item \code{df} must contain the variables specified by
\code{arm_var}, \code{id_var}, \code{soc_var} and \code{saffl_var}.
\item \code{return_ard} set to \code{TRUE} (default) or \code{FALSE}; whether the intermediate ARD object should be returned.
}
}
\examples{
adsl <- random.cdisc.data::cadsl
adae <- random.cdisc.data::cadae

tbl <- make_table_35(df = adae, denominator = adsl)
tbl

}
\seealso{
\code{\link{tbl_make_table_35}}
}
