\name{MSImageProcess-class}
\docType{class}

\alias{class:MSImageProcess}
\alias{MSImageProcess}
\alias{MSImageProcess-class}

\alias{show,MSImageProcess-method}
\alias{combine,MSImageProcess,MSImageProcess-method}
\alias{initialize,MSImageProcess-method}
\alias{files}
\alias{files,MSImageProcess-method}
\alias{files<-}
\alias{files<-,MSImageProcess-method}
\alias{normalization}
\alias{normalization,MSImageProcess-method}
\alias{normalization<-}
\alias{normalization<-,MSImageProcess-method}
\alias{smoothing}
\alias{smoothing,MSImageProcess-method}
\alias{smoothing<-}
\alias{smoothing<-,MSImageProcess-method}
\alias{baselineReduction}
\alias{baselineReduction,MSImageProcess-method}
\alias{baselineReduction<-}
\alias{baselineReduction<-,MSImageProcess-method}
\alias{spectrumRepresentation}
\alias{spectrumRepresentation,MSImageProcess-method}
\alias{spectrumRepresentation<-}
\alias{spectrumRepresentation<-,MSImageProcess-method}
\alias{peakPicking}
\alias{peakPicking,MSImageProcess-method}
\alias{peakPicking<-}
\alias{peakPicking<-,MSImageProcess-method}
\alias{centroided}
\alias{centroided,MSImageProcess-method}
\alias{centroided<-}
\alias{centroided<-,MSImageProcess-method}
\alias{prochistory}
\alias{prochistory,MSImageProcess-method}
\alias{prochistory<-}
\alias{prochistory<-,MSImageProcess-method}
\alias{prochistory<-,MSImageProcess,character-method}
\alias{prochistory<-,MSImageProcess,list-method}

\title{Class Containing Mass Spectral Pre-Processing Information}

\description{
    A class containing information about mass spectral pre-processsing operations. These should not usually be set by the user, and are automatically updated when processing methods are applied.
}

\section{Slots}{
    \describe{
        \item{\code{files}:}{Object of class \code{character} storing the file paths to the raw data files used to create the dataset.}

        \item{\code{normalization}:}{Object of class \code{character} describing any normalization applied to the dataset.}
        
        \item{\code{smoothing}:}{Object of class \code{character} describing any smoothing applied to the dataset.}
        
        \item{\code{baselineReduction}:}{Object of class \code{character} describing baseline correction applied to the dataset.}

        \item{\code{spectrumRepresentation}:}{Object of class \code{character} describing the spectrum type (profile or centroid).}

        \item{\code{peakPicking}:}{Object of class \code{character} describing the peak picking applied to the dataset (area or height).}

        \item{\code{centroided}:}{Object of class \code{logical} describing whether the data have been centroided.}

        \item{\code{history}:}{Object of class \code{list} containing specific information about the function calls applied to the \code{\linkS4class{MSImageSet} object} to produce the current instance and their parameters.}

        \item{\code{CardinalVersion}:}{Object of class \code{character} indicating the version of Cardinal.}

        \item{\code{.__classVersion__}:}{Object of class \code{Versions} indicating the version of the \code{MSImageProcess} instance. Intended for developer use.}
    }
}

\section{Extends}{
   Class \code{\linkS4class{Versioned}}, directly.
}

\section{Creating Objects}{
    \code{MSImageProcess} instances can be created through \code{new("MSImageProcess")}. In general, instances should not be created by the user, but are automatically generated by processing methods applied to \code{\linkS4class{MSImageSet}} objects.
}

\section{Methods}{
    Class-specific methods:
    \describe{
        \item{\code{prochistory(object), prochistory(object) <- value}:}{Accessor and setter for the history of methods applied to the experiment dataset.}

        \item{\code{files(object), files(object) <- value}:}{Accessor and setter function for \code{files}.}

        \item{\code{normalization(object), normalization(object) <- value}:}{Accessor and setter function for \code{normalization}.}
        
        \item{\code{smoothing(object), smoothing(object) <- value}:}{Accessor and setter function for \code{smoothing}.}
        
        \item{\code{baselineReduction(object), baselineReduction(object) <- value}:}{Accessor and setter function for \code{baselineReduction}.}

        \item{\code{spectrumRepresentation(object), spectrumRepresentation(object) <- value}:}{Accessor and setter function for \code{spectrumRepresentation}.}

        \item{\code{peakPicking(object), peakPicking(object) <- value}:}{Accessor and setter function for \code{peakPicking}.}

        \item{\code{centroided(object), centroided(object) <- value}:}{Accessor and setter function for \code{centroided}.}
    }

    Standard generic methods:
    \describe{
        \item{\code{show}:}{Displays object content.}

        \item{\code{combine(x, y, ...)}:}{Combine two or more \code{MSImageProcess} objects.}
    }
}

\author{Kyle D. Bemis}

\seealso{
    \code{\linkS4class{MSImageSet}}
}

\examples{
showClass("MSImageProcess")
}

\keyword{classes}
