\name{reexports}
\alias{Cardinal-reexports}
\docType{import}

\alias{\%>\%}

\alias{groups}
\alias{group_by}
\alias{ungroup}

\alias{viridis}
\alias{cividis}
\alias{magma}
\alias{inferno}
\alias{plasma}

\title{Objects exported from other packages}

\description{
  These objects are imported from other packages and have been re-exported by Cardinal for user convenience.

  \describe{
        \item{\code{maggritr}:}{\code{\%>\%}}

		\item{\code{dplyr}:}{\code{group_by}, \code{ungroup}, \code{groups}}

        \item{\code{viridisLite}:}{\code{viridis}, \code{cividis}, \code{magma}, \code{inferno}, \code{plasma}}
    }
}
