\name{reexports}
\alias{reexports}

\alias{SnowfastParam}

\alias{cpal}
\alias{cpals}
\alias{dpal}
\alias{dpals}

\alias{as_facets}
\alias{as_layers}

\title{Re-exported objects from Cardinal}

\description{
	These functions are re-exported from Cardinal for user convenience. Please see the documentation in their original packages.
}
