\name{reexports}
\alias{Cardinal-reexports}
\docType{import}

\alias{\%>\%}

\title{Objects exported from other packages}

\description{
  These objects are imported from other packages and have been re-exported by Cardinal for user convenience.

  \describe{
        \item{\code{maggritr}:}{\code{\%>\%}}
    }
}
