% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_cols_by_arm}
\alias{split_cols_by_arm}
\title{Split Columns by Arm Variable}
\usage{
split_cols_by_arm(lyt, arm_var = "ARM", lbl_overall = NULL, risk_diff = NULL)
}
\arguments{
\item{lyt}{(\code{rtables} layout object)\cr Layout object pre-data used for tabulation.}

\item{arm_var}{(\code{character})\cr arm variable used to split table into columns.}

\item{lbl_overall}{(\code{character})\cr if specified, an overall column will be added to the table with
the given value as the column label.}

\item{risk_diff}{(named \code{list})\cr list of settings to apply to add a risk difference column to the table.
Defaults to \code{NULL} (no risk difference column added). See \code{\link[tern:add_riskdiff]{tern::add_riskdiff()}} for more details. List should
contain the following elements:
\itemize{
\item \code{arm_x}: (required) the name of reference arm.
\item \code{arm_y}: (required) the name of the arm to compare to the reference arm.
\item \code{col_label}: (optional) label to use for the risk difference column. Defaults to
\code{"Risk Difference (\%) (95\% CI)"}.
\item \code{pct}: (optional) whether the output should be returned as percentages. Defaults to \code{TRUE}.
}}
}
\value{
An \code{rtables} \code{PreDataTableLayouts} object suitable for passing to further layout functions, and
to \code{build_table}.
}
\description{
Splits \code{rtables} layout into columns by arm variable. Appends an "overall" column as the
rightmost column if \code{lbl_overall} is supplied.
}
\examples{
library(magrittr)

adsl <- scda::synthetic_cdisc_dataset("rcd_2022_10_13", "adsl")

lyt <- rtables::basic_table() \%>\% split_cols_by_arm(lbl_overall = "All Arms")
rtables::build_table(lyt, df = adsl)

}
